@@ -1,47 +1,163 @@
 {
   "name": "ganache-core",
-  "version": "2.0.0",
+  "version": "2.7.0-beta.0",
   "main": "./index.js",
+  "types": "./typings/index.d.ts",
+  "engines": {
+    "node": ">=8.9.0"
+  },
   "directories": {
     "lib": "./lib"
   },
   "scripts": {
-    "test": "mocha"
+    "_mocha": "mocha --check-leaks --recursive --globals _scratch --opts ./test/.mocharc",
+    "nyc_mocha": "nyc mocha --check-leaks --recursive --globals _scratch --opts ./test/.mocharc",
+    "_lint": "eslint --ignore-path .gitignore .",
+    "build": "webpack-cli --config ./webpack/node/core.webpack.config.js",
+    "build-web": "webpack-cli --config ./webpack/web-experimental/core.webpack.config.js",
+    "format": "prettier --write \"{lib,perf,test}/**/*.js\" && eslint --fix --ignore-path .gitignore .",
+    "prepublishOnly": "npm run test && npm run build && npm run test-build && cp npm-shrinkwrap.json npm-shrinkwrap.json.bak && rm -rf node_modules/web3-providers-ws/node_modules/websocket/.git && npm prune --only=prod",
+    "postpublish": "rm npm-shrinkwrap.json && mv npm-shrinkwrap.json.bak npm-shrinkwrap.json && npm ci",
+    "test": "npm run _lint && npm run nyc_mocha",
+    "test-build": "cross-env TEST_BUILD=node npm run _mocha",
+    "test-web-experimental": "cross-env TEST_BUILD=web-experimental npm run _mocha",
+    "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "dependencies": {
-    "async": "~1.5.0",
-    "bip39": "~2.4.0",
-    "cachedown": "^1.0.0",
-    "chai": "^3.5.0",
-    "clone": "^2.1.1",
-    "ethereumjs-account": "~2.0.4",
-    "ethereumjs-block": "~1.2.2",
-    "ethereumjs-tx": "^1.3.0",
-    "ethereumjs-util": "~5.1.0",
-    "ethereumjs-vm": "https://github.com/sc-forks/ethereumjs-vm-sc.git#coverage",
-    "ethereumjs-wallet": "~0.6.0",
-    "fake-merkle-patricia-tree": "~1.0.1",
-    "heap": "~0.2.6",
-    "js-scrypt": "^0.2.0",
-    "level-sublevel": "^6.6.1",
-    "levelup": "^1.1.0",
-    "localstorage-down": "^0.6.7",
-    "merkle-patricia-tree": "^2.2.0",
-    "mocha": "~3.3.0",
-    "on-build-webpack": "^0.1.0",
-    "prepend-file": "^1.3.1",
-    "seedrandom": "~2.4.2",
-    "shebang-loader": "0.0.1",
-    "solc": "0.4.18",
-    "temp": "^0.8.3",
-    "tmp": "0.0.31",
-    "web3": "^0.19.1",
-    "web3-provider-engine": "https://github.com/sc-forks/provider-engine-sc.git#8.1.19",
-    "webpack": "^2.2.1",
-    "yargs": "^7.0.2"
+    "abstract-leveldown": "3.0.0",
+    "async": "2.6.2",
+    "bip39": "2.5.0",
+    "cachedown": "1.0.0",
+    "clone": "2.1.2",
+    "debug": "3.2.6",
+    "encoding-down": "5.0.4",
+    "eth-sig-util": "2.2.0",
+    "ethereumjs-abi": "0.6.7",
+    "ethereumjs-account": "3.0.0",
+    "ethereumjs-block": "2.2.0",
+    "ethereumjs-tx": "1.3.7",
+    "ethereumjs-util": "6.1.0",
+    "ethereumjs-vm": "3.0.0",
+    "heap": "0.2.6",
+    "level-sublevel": "6.6.4",
+    "levelup": "3.1.1",
+    "lodash": "4.17.14",
+    "merkle-patricia-tree": "2.3.2",
+    "seedrandom": "3.0.1",
+    "source-map-support": "0.5.12",
+    "tmp": "0.1.0",
+    "web3-provider-engine": "14.2.0",
+    "websocket": "1.0.29"
+  },
+  "devDependencies": {
+    "@types/web3": "^1.0.19",
+    "assert-match": "^1.1.1",
+    "browserfs": "1.4.3",
+    "coveralls": "^3.0.4",
+    "cross-env": "5.2.0",
+    "eslint": "5.7.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-standard": "4.0.0",
+    "ethereumjs-wallet": "0.6.3",
+    "ethers": "4.0.26",
+    "husky": "^1.1.3",
+    "js-scrypt": "0.2.0",
+    "lint-staged": "^8.0.4",
+    "memdown": "^1.3.1",
+    "mocha": "5.2.0",
+    "mocha-lcov-reporter": "^1.3.0",
+    "number-to-bn": "1.7.0",
+    "nyc": "^14.1.1",
+    "pify": "4.0.0",
+    "portfinder": "^1.0.18",
+    "prettier": "^1.18.2",
+    "request": "^2.88.0",
+    "semver": "5.6.0",
+    "solc": "0.5.10",
+    "temp": "0.8.3",
+    "terser-webpack-plugin": "1.3.0",
+    "web3": "1.0.0-beta.35",
+    "webpack": "4.17.1",
+    "webpack-bundle-size-analyzer": "2.7.0",
+    "webpack-cli": "3.1.0"
+  },
+  "optionalDependencies": {
+    "ethereumjs-wallet": "0.6.3",
+    "web3": "1.0.0-beta.35"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/trufflesuite/ganache-core"
+  },
+  "license": "MIT",
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.js": [
+      "prettier --write",
+      "eslint --fix --ignore-path .gitignore",
+      "git add"
+    ]
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true,
+      "es6": true
+    },
+    "parserOptions": {
+      "ecmaVersion": 8,
+      "sourceType": "module",
+      "ecmaFeatures": {
+        "impliedStrict": true
+      }
+    },
+    "extends": "standard",
+    "rules": {
+      "brace-style": [
+        "error",
+        "1tbs",
+        {
+          "allowSingleLine": false
+        }
+      ],
+      "curly": [
+        "error",
+        "all"
+      ],
+      "max-len": [
+        "error",
+        120,
+        {
+          "ignoreRegExpLiterals": true
+        }
+      ],
+      "quotes": [
+        "error",
+        "double"
+      ],
+      "semi": [
+        "error",
+        "always"
+      ],
+      "space-before-function-paren": [
+        "error",
+        "never"
+      ],
+      "wrap-iife": [
+        "error",
+        "outside"
+      ]
+    }
+  },
+  "prettier": {
+    "printWidth": 120,
+    "arrowParens": "always"
   }
 }
