@@ -1,26 +1,27 @@
-var inherits = require("util").inherits;
 var to = require("./utils/to.js");
-var Account = require('ethereumjs-account');
-var Block = require('ethereumjs-block');
+var Account = require("ethereumjs-account").default;
+var Block = require("ethereumjs-block");
 var Log = require("./utils/log");
 var Receipt = require("./utils/receipt");
-var VM = require('ethereumjs-vm');
+var VM = require("ethereumjs-vm");
 var RuntimeError = require("./utils/runtimeerror");
 var Trie = require("merkle-patricia-tree");
-var Web3 = require("web3");
 var utils = require("ethereumjs-util");
-var async = require('async');
+var async = require("async");
 var Heap = require("heap");
 var Database = require("./database");
-var path = require("path");
-var clone = require("clone");
+var EventEmitter = require("events");
+var estimateGas = require("./utils/gasEstimation");
+var _ = require("lodash");
+var promisify = require("util").promisify;
+const BN = utils.BN;
 
 function BlockchainDouble(options) {
   var self = this;
+  EventEmitter.apply(self);
 
-  options = options || {};
+  this.options = options = this._applyDefaultOptions(options || {});
 
-  this.options = options;
   this.logger = options.logger || console;
 
   this.data = new Database(options);
@@ -31,21 +32,43 @@ function BlockchainDouble(options) {
 
   this.pending_transactions = [];
 
-  // Homestead block gas limit is 4712388 == 0x47E7C4
-  // Default transaction gas limit is 90000 == 0x15f90
-  this.blockGasLimit = options.gasLimit || "0xfffffffffff";
-  this.defaultTransactionGasLimit = '0xfffffff';
+  // updated periodically to keep up with the times
+  this.blockGasLimit = options.gasLimit;
+  this.defaultTransactionGasLimit = options.defaultTransactionGasLimit;
   this.timeAdjustment = 0;
+}
+
+const defaultOptions = {
+  gasLimit: "0x6691b7",
+  defaultTransactionGasLimit: "0x15f90",
+  time: null,
+  debug: false,
+  hardfork: "petersburg",
+  allowUnlimitedContractSize: false
+};
+
+// inheritence w/ prototype chaining
+BlockchainDouble.prototype = Object.create(EventEmitter.prototype);
+BlockchainDouble.prototype.constructor = BlockchainDouble;
+
+BlockchainDouble.prototype._applyDefaultOptions = function(options) {
+  // We want this function to mutate the options object so that we can report
+  // our settings back to our consumer application (e.g., ganache)
+  return _.merge(options, defaultOptions, Object.assign({}, options));
 };
 
 BlockchainDouble.prototype.initialize = function(accounts, callback) {
   var self = this;
 
   this.data.initialize(function(err) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     self.latestBlock(function(err, block) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       var options = self.options;
 
@@ -59,36 +82,7 @@ BlockchainDouble.prototype.initialize = function(accounts, callback) {
       // forked blockchain without going through the other setup is a little gross.
       self.stateTrie = self.createStateTrie(self.data.trie_db, root);
 
-      self.vm = options.vm || new VM({
-        state: self.stateTrie,
-        blockchain: {
-          // EthereumJS VM needs a blockchain object in order to get block information.
-          // When calling getBlock() it will pass a number that's of a Buffer type.
-          // Unfortunately, it uses a 64-character buffer (when converted to hex) to
-          // represent block numbers as well as block hashes. Since it's very unlikely
-          // any block number will get higher than the maximum safe Javascript integer,
-          // we can convert this buffer to a number ahead of time before calling our
-          // own getBlock(). If the conversion succeeds, we have a block number.
-          // If it doesn't, we have a block hash. (Note: Our implementation accepts both.)
-          getBlock: function(number, done) {
-            try {
-              number = to.number(number)
-            } catch (e) {
-              // Do nothing; must be a block hash. 
-            }
-
-            self.getBlock(number, done)
-          } 
-        },
-        enableHomestead: true,
-        activatePrecompiles: true
-      });
-
-      if (options.debug == true) {
-        self.vm.on('step', function(info){
-          self.logger.log(info.opcode.name)
-        });
-      }
+      self.vm = options.vm || self.createVMFromStateTrie(self.stateTrie, true);
 
       if (options.time) {
         self.setTime(options.time);
@@ -97,30 +91,91 @@ BlockchainDouble.prototype.initialize = function(accounts, callback) {
       // If we already have a block, then that means there's an existing chain.
       // Don't create a genesis block.
       if (block) {
+        self.emit("block", block);
         return callback();
       }
 
       self.createGenesisBlock(function(err, block) {
-        if (err) return callback(err);
+        if (err) {
+          return callback(err);
+        }
 
         accounts = accounts || [];
 
-        async.eachSeries(accounts, function(account_data, finished) {
-          self.putAccount(account_data.account, account_data.address, finished);
-        }, function(err) {
-          if (err) return callback(err);
+        self.vm.stateManager.checkpoint(() => {
+          async.eachSeries(
+            accounts,
+            function(accountData, finished) {
+              self.vm.stateManager.putAccount(utils.toBuffer(accountData.address), accountData.account, finished);
+            },
+            function(err) {
+              if (err) {
+                return callback(err);
+              }
 
-          // Create first block
-          self.putBlock(block, [], [], callback);
+              self.vm.stateManager.commit(() => {
+                // Create first block
+                self.putBlock(block, [], [], callback);
+              });
+            }
+          );
         });
       });
     });
-
   });
 };
 
+BlockchainDouble.prototype.createVMFromStateTrie = function(state, activatePrecompiles) {
+  const self = this;
+  const vm = new VM({
+    state: state,
+    blockchain: {
+      // EthereumJS VM needs a blockchain object in order to get block information.
+      // When calling getBlock() it will pass a number that's of a Buffer type.
+      // Unfortunately, it uses a 64-character buffer (when converted to hex) to
+      // represent block numbers as well as block hashes. Since it's very unlikely
+      // any block number will get higher than the maximum safe Javascript integer,
+      // we can convert this buffer to a number ahead of time before calling our
+      // own getBlock(). If the conversion succeeds, we have a block number.
+      // If it doesn't, we have a block hash. (Note: Our implementation accepts both.)
+      getBlock: function(number, done) {
+        try {
+          number = to.number(number);
+        } catch (e) {
+          // Do nothing; must be a block hash.
+        }
+
+        self.getBlock(number, done);
+      }
+    },
+    activatePrecompiles: activatePrecompiles || false,
+    hardfork: self.options.hardfork,
+    allowUnlimitedContractSize: self.options.allowUnlimitedContractSize
+  });
+
+  if (self.options.debug === true) {
+    // log executed opcodes, including args as hex
+    vm.on("step", function(info) {
+      var name = info.opcode.name;
+      var argsNum = info.opcode.in;
+      if (argsNum) {
+        var args = info.stack
+          .slice(-argsNum)
+          .map((arg) => to.hex(arg))
+          .join(" ");
+
+        self.logger.log(`${name} ${args}`);
+      } else {
+        self.logger.log(name);
+      }
+    });
+  }
+
+  return vm;
+};
+
 BlockchainDouble.prototype.createStateTrie = function(db, root) {
-  return new Trie(this.data.trie_db, root);
+  return new Trie(db, root);
 };
 
 // Overrideable so other implementations (forking) can edit it.
@@ -130,14 +185,16 @@ BlockchainDouble.prototype.createGenesisBlock = function(callback) {
 
 BlockchainDouble.prototype.latestBlock = function(callback) {
   this.data.blocks.last(function(err, last) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     callback(null, last);
   });
-}
+};
 
 // number accepts number (integer, hex) or tag (e.g., "latest")
 BlockchainDouble.prototype.getEffectiveBlockNumber = function(number, callback) {
-  if (typeof number != "string") {
+  if (typeof number !== "string") {
     number = to.hex(number);
   }
 
@@ -145,9 +202,9 @@ BlockchainDouble.prototype.getEffectiveBlockNumber = function(number, callback) 
   if (number.indexOf("0x") >= 0) {
     return callback(null, to.number(number));
   } else {
-    if (number == "latest" || number == "pending") {
+    if (number === "latest" || number === "pending") {
       return this.getHeight(callback);
-    } else if (number == "earliest") {
+    } else if (number === "earliest") {
       return callback(null, 0);
     }
   }
@@ -158,7 +215,7 @@ BlockchainDouble.prototype.getEffectiveBlockNumber = function(number, callback) 
 BlockchainDouble.prototype.getBlock = function(number, callback) {
   var self = this;
 
-  if (typeof number != "string") {
+  if (typeof number !== "string") {
     number = to.hex(number);
   }
 
@@ -169,7 +226,9 @@ BlockchainDouble.prototype.getBlock = function(number, callback) {
     // block hash
     if (hash.length > 40) {
       this.data.blockHashes.get(to.hex(hash), function(err, blockIndex) {
-        if (err) return callback(err);
+        if (err) {
+          return callback(err);
+        }
         return self.data.blocks.get(blockIndex, callback);
       });
     } else {
@@ -177,10 +236,12 @@ BlockchainDouble.prototype.getBlock = function(number, callback) {
       return this.data.blocks.get(to.number(hash), callback);
     }
   } else {
-    if (number == "latest" || number == "pending") {
+    if (number === "latest" || number === "pending") {
       return this.latestBlock(callback);
-    } else if (number == "earliest") {
+    } else if (number === "earliest") {
       return this.data.blocks.first(callback);
+    } else {
+      process.nextTick(callback, new Error("Invalid `blockNumber`: \"" + number + "\""));
     }
   }
 };
@@ -192,24 +253,30 @@ BlockchainDouble.prototype.putBlock = function(block, logs, receipts, callback) 
   block.header.stateRoot = this.stateTrie.root;
 
   this.data.blocks.length(function(err, length) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     var requests = [
       self.data.blocks.push.bind(self.data.blocks, block),
       self.data.blockLogs.push.bind(self.data.blockLogs, logs),
-      self.data.blockHashes.set.bind(self.data.blockHashes, to.hex(block.hash()), length),
+      self.data.blockHashes.set.bind(self.data.blockHashes, to.hex(block.hash()), length)
     ];
 
     block.transactions.forEach(function(tx, index) {
-      var tx_hash = to.hex(tx.hash());
-
+      var txHash = to.hex(tx.hash());
       requests.push(
-        self.data.transactions.set.bind(self.data.transactions, tx_hash, tx),
-        self.data.transactionReceipts.set.bind(self.data.transactionReceipts, tx_hash, receipts[index])
+        self.data.transactions.set.bind(self.data.transactions, txHash, tx),
+        self.data.transactionReceipts.set.bind(self.data.transactionReceipts, txHash, receipts[index])
       );
     });
 
-    async.parallel(requests, callback);
+    async.parallel(requests, (err, result) => {
+      if (!err) {
+        self.emit("block", block);
+      }
+      callback(err, result);
+    });
   });
 };
 
@@ -217,18 +284,22 @@ BlockchainDouble.prototype.popBlock = function(callback) {
   var self = this;
 
   this.data.blocks.last(function(err, block) {
-    if (err) return callback(err);
-    if (block == null) return callback(null, null);
+    if (err) {
+      return callback(err);
+    }
+    if (block == null) {
+      return callback(null, null);
+    }
 
     var requests = [];
     var blockHash = to.hex(block.hash());
 
-    block.transactions.forEach(function(tx, index) {
-      var tx_hash = to.hex(tx.hash());
+    block.transactions.forEach(function(tx) {
+      var txHash = to.hex(tx.hash());
 
       requests.push(
-        self.data.transactions.del.bind(self.data.transactions, tx_hash),
-        self.data.transactionReceipts.del.bind(self.data.transactionReceipts, tx_hash)
+        self.data.transactions.del.bind(self.data.transactions, txHash),
+        self.data.transactionReceipts.del.bind(self.data.transactionReceipts, txHash)
       );
     });
 
@@ -239,15 +310,31 @@ BlockchainDouble.prototype.popBlock = function(callback) {
     );
 
     async.series(requests, function(err) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       // Set the root to the last available, which will "roll back" to the previous
       // moment in time. Note that all the old data is still in the db, but it's now just junk data.
       self.data.blocks.last(function(err, newLastBlock) {
-        if (err) return callback(err);
-        self.stateTrie.root = newLastBlock.header.stateRoot;
-        // Remember: Return block we popped off.
-        callback(null, block);
+        if (err) {
+          return callback(err);
+        }
+        // using setStateRoot because in the future it will automatically take care
+        // of clearing the cache for us.
+        // note setStateRoot checks for checkpoints, and if there are any, it will fail.
+        // At time of writing this comment, the only time there could be a checkpoint
+        // is in the middle of a vm.runBlock call. Once asyncRequestProcessing is reenabled
+        // this will likely cause problems.
+        self.vm.stateManager.setStateRoot(newLastBlock.header.stateRoot, function(err) {
+          // remove this `._storageTries = {}` statement once https://github.com/ethereumjs/ethereumjs-vm/pull/445
+          // is in a release, probably in ejs-vm 2.6.1
+          if (!err) {
+            self.vm.stateManager._storageTries = {};
+          }
+          // Remember: Return block we popped off.
+          callback(err, block);
+        });
       });
     });
   });
@@ -255,18 +342,6 @@ BlockchainDouble.prototype.popBlock = function(callback) {
 
 BlockchainDouble.prototype.clearPendingTransactions = function() {
   this.pending_transactions = [];
-};
-
-BlockchainDouble.prototype.putAccount = function(account, address, callback) {
-  var self = this;
-
-  address = utils.toBuffer(address);
-
-  this.vm.stateManager._putAccount(address, account, function(err) {
-    if (err) return callback(err);
-
-    self.vm.stateManager.cache.flush(callback);
-  });
 };
 
 /**
@@ -279,12 +354,16 @@ BlockchainDouble.prototype.putAccount = function(account, address, callback) {
  * @param  {Function} callback Callback function called after block is created
  * @return Block              The block created.
  */
-BlockchainDouble.prototype.createBlock = function(parent, callback) {
+BlockchainDouble.prototype.createBlock = function(parent, emulateParent, callback) {
   var self = this;
 
-  if (typeof parent == "function") {
+  if (typeof parent === "function") {
     callback = parent;
     parent = null;
+    emulateParent = false;
+  } else if (typeof emulateParent === "function") {
+    callback = emulateParent;
+    emulateParent = false;
   }
 
   var block = new Block();
@@ -295,23 +374,31 @@ BlockchainDouble.prototype.createBlock = function(parent, callback) {
     } else {
       self.latestBlock(callback);
     }
-  };
+  }
 
   getParent(function(err, parent) {
-    if (err) return callback(err);
-
-    var parentNumber = parent != null ? to.number(parent.header.number) : -1;
+    if (err) {
+      return callback(err);
+    }
 
     block.header.gasLimit = self.blockGasLimit;
 
-    // Ensure we have the right block number for the VM.
-    block.header.number = to.hex(parentNumber + 1);
+    if (parent != null && emulateParent) {
+      block.header.number = parent.header.number;
+      block.header.timestamp = parent.header.timestamp;
+      block.header.parentHash = parent.header.parentHash;
+    } else {
+      var parentNumber = parent != null ? to.number(parent.header.number) : -1;
 
-    // Set the timestamp before processing txs
-    block.header.timestamp = to.hex(self.currentTime());
+      // Ensure we have the right block number for the VM.
+      block.header.number = to.hex(parentNumber + 1);
 
-    if (parent != null) {
-      block.header.parentHash = to.hex(parent.hash());
+      // Set the timestamp before processing txs
+      block.header.timestamp = to.hex(self.currentTime());
+
+      if (parent != null) {
+        block.header.parentHash = to.hex(parent.hash());
+      }
     }
 
     callback(null, block);
@@ -320,29 +407,33 @@ BlockchainDouble.prototype.createBlock = function(parent, callback) {
 
 BlockchainDouble.prototype.getQueuedNonce = function(address, callback) {
   var nonce = null;
-
+  var addressBuffer = to.buffer(address);
   this.pending_transactions.forEach(function(tx) {
-    //tx.from and address are buffers, so cannot simply do
-    //tx.from==address
-    if (to.hex(tx.from) != to.hex(address)) return;
+    if (!tx.from.equals(addressBuffer)) {
+      return;
+    }
 
-    var pending_nonce = to.number(tx.nonce);
-    //If this is the first queued nonce for this address we found,
-    //or it's higher than the previous highest, note it.
-    if (nonce===null || pending_nonce > nonce) {
-      nonce = pending_nonce;
+    var pendingNonce = new BN(tx.nonce);
+    // If this is the first queued nonce for this address we found,
+    // or it's higher than the previous highest, note it.
+    if (nonce === null || pendingNonce.gt(nonce)) {
+      nonce = pendingNonce;
     }
   });
 
-  //If we found a queued transaction nonce, return one higher
-  //than the highest we found
-  if (nonce!=null) return callback(null, nonce+1);
-
-  this.stateTrie.get(address, function(err, val) {
-    if (err) return callback(err);
+  // If we found a queued transaction nonce, return one higher
+  // than the highest we found
+  if (nonce != null) {
+    return callback(null, nonce.iaddn(1).toArrayLike(Buffer));
+  }
+  this.stateTrie.get(addressBuffer, function(err, val) {
+    if (err) {
+      return callback(err);
+    }
 
     var account = new Account(val);
-    callback(null, account.nonce);
+    // nonces are initiallized as an empty buffer, which isn't what we want.
+    callback(null, account.nonce.length === 0 ? Buffer.from([0]) : account.nonce);
   });
 };
 
@@ -356,217 +447,256 @@ BlockchainDouble.prototype.sortByPriceAndNonce = function() {
   // https://github.com/ethereum/go-ethereum/blob/290e851f57f5d27a1d5f0f7ad784c836e017c337/core/types/transaction.go
   var self = this;
   var sortedByNonce = {};
-  for (idx in self.pending_transactions){
-    var tx = self.pending_transactions[idx]
-    if (!sortedByNonce[to.hex(tx.from)]){
-      sortedByNonce[to.hex(tx.from)] = [tx];
+
+  self.pending_transactions.forEach((tx) => {
+    const from = tx.from.toString("hex");
+    const arr = sortedByNonce[from];
+    if (arr) {
+      arr.push(tx);
     } else {
-      Array.prototype.push.apply(sortedByNonce[to.hex(tx.from)], [tx]);
+      sortedByNonce[from] = [tx];
     }
-  }
-  var priceSort = function(a,b){
-    return parseInt(to.hex(b.gasPrice),16)-parseInt(to.hex(a.gasPrice),16);
-  }
-  var nonceSort = function(a,b){
-    return parseInt(to.hex(a.nonce),16) - parseInt(to.hex(b.nonce),16)
-  }
+  });
+
+  var priceSort = function(a, b) {
+    return parseInt(to.hex(b.gasPrice), 16) - parseInt(to.hex(a.gasPrice), 16);
+  };
+  var nonceSort = function(a, b) {
+    return parseInt(to.hex(a.nonce), 16) - parseInt(to.hex(b.nonce), 16);
+  };
 
   // Now sort each address by nonce
-  for (address in sortedByNonce){
-    Array.prototype.sort.apply(sortedByNonce[address], [nonceSort])
-  }
+  Object.keys(sortedByNonce).forEach((address) => {
+    sortedByNonce[address].sort(nonceSort);
+  });
 
   // Initialise a heap, sorted by price, for the head transaction from each account.
   var heap = new Heap(priceSort);
-  for (address in sortedByNonce){
-    heap.push(sortedByNonce[address][0]);
-    //Remove the transaction from sortedByNonce
-    sortedByNonce[address].splice(0,1);
-  }
+  Object.keys(sortedByNonce).forEach((address) => {
+    heap.push(sortedByNonce[address].shift());
+  });
 
   // Now reorder our transactions. Compare the next transactions from each account, and choose
   // the one with the highest gas price.
-  sorted_transactions = [];
-  while (heap.size()>0){
-    best = heap.pop();
-    if (sortedByNonce[to.hex(best.from)].length>0){
-      //Push on the next transaction from this account
-      heap.push(sortedByNonce[address][0]);
-      sortedByNonce[address].splice(0,1);
+  const sortedTransactions = [];
+  while (heap.size() > 0) {
+    const best = heap.pop();
+    let address = best.from.toString("hex");
+    if (sortedByNonce[address].length > 0) {
+      // Push on the next transaction from this account
+      heap.push(sortedByNonce[address].shift());
     }
-    Array.prototype.push.apply(sorted_transactions, [best]);
+    sortedTransactions.push(best);
   }
-  self.pending_transactions = sorted_transactions;
+  self.pending_transactions = sortedTransactions;
 };
 
-BlockchainDouble.prototype.processCall = function(tx, callback) {
-    var self = this;
-
+BlockchainDouble.prototype.readyCall = function(tx, emulateParent, blockNumber, callback) {
+  const readyCall = (tx, err, parentBlock) => {
+    if (err) {
+      return callback(err);
+    }
     // create a fake block with this fake transaction
-    var block = new Block();
-    block.transactions.push(tx);
-
-    self.latestBlock(function (err, latestBlock) {
+    this.createBlock(parentBlock, emulateParent, (err, newBlock) => {
       if (err) {
         return callback(err);
       }
 
-      // we need to fake out the block number as if it were the current block
-      // as otherwise the VM will report that it's operating on a block which
-      // doesn't exist
-      block.header = clone(latestBlock.header, false);
+      newBlock.transactions.push(tx);
 
-      // We checkpoint here for speed. We want all state trie reads/writes to happen in memory,
-      // and the final output be flushed to the database at the end of transaction processing.
-      self.vm.stateManager.checkpoint();
+      // gas estimates and eth_calls aren't subject to block gas limits
+      newBlock.header.gasLimit = tx.gasLimit;
 
-      var cleanUpAndReturn = function (err, result) {
-        self.vm.stateManager.revert(function (e) {
-          callback(err || e, result);
-        });
-      };
-
-      var runArgs = {
+      const runArgs = {
         tx: tx,
-        block: block,
+        block: newBlock,
         skipBalance: true,
         skipNonce: true
       };
-
-      self.vm.runTx(runArgs, function (vmerr, result) {
-        // This is a check that has been in there for awhile. I'm unsure if it's required, but it can't hurt.
-        if (vmerr && vmerr instanceof Error == false) {
-          vmerr = new Error("VM error: " + vmerr);
-        }
-
-        // If we're given an error back directly, it's worse than a runtime error. Expose it and get out.
-        if (vmerr) return cleanUpAndReturn(vmerr);
-
-        // If no error, check for a runtime error. This can return null if no runtime error.
-        vmerr = RuntimeError.fromResults([tx], { results: [result] });
-
-        cleanUpAndReturn(vmerr, result);
-      });
+      const stateTrie = this.createStateTrie(this.data.trie_db, parentBlock.header.stateRoot);
+      const vm = this.createVMFromStateTrie(stateTrie);
+      callback(null, vm, runArgs);
     });
-}
+  };
+  // Delegate block selection
+  if (blockNumber === "latest") {
+    this.latestBlock(readyCall.bind(null, tx));
+  } else {
+    this.getBlock(blockNumber, readyCall.bind(null, tx));
+  }
+};
+
+BlockchainDouble.prototype.processCall = function(tx, blockNumber, callback) {
+  this.readyCall(tx, true, blockNumber, (err, vm, runArgs) => {
+    if (err) {
+      callback(err);
+      return;
+    }
+
+    vm.runTx(runArgs, function(vmerr, result) {
+      // This is a check that has been in there for awhile. I'm unsure if it's required, but it can't hurt.
+      if (vmerr && vmerr instanceof Error === false) {
+        vmerr = new Error("VM error: " + vmerr);
+      }
+
+      // If we're given an error back directly, it's worse than a runtime error. Expose it and get out.
+      if (vmerr) {
+        return callback(vmerr, err);
+      }
+
+      // If no error, check for a runtime error. This can return null if no runtime error.
+      vmerr = RuntimeError.fromResults([tx], { results: [result] });
+
+      callback(vmerr, result);
+    });
+  });
+};
+
+BlockchainDouble.prototype.estimateGas = function(tx, blockNumber, callback) {
+  this.readyCall(tx, false, blockNumber, (err, vm, runArgs) => {
+    if (err) {
+      callback(err);
+      return;
+    }
+
+    estimateGas(vm, runArgs, err, callback);
+  });
+};
 
 /**
  * processBlock
  *
  * Process the passed in block and included transactions
  *
+ * @param  {VM} vm             the vm to use when running the block
  * @param  {Block} block       block to process
- * @param  {Boolean} commit    Whether or not changes should be committed to the state trie and the block appended to the end of the chain.
+ * @param  {Boolean} commit    Whether or not changes should be committed to the state
+ * trie and the block appended to the end of the chain.
  * @param  {Function} callback Callback function when transaction processing is completed.
  * @return [type]              [description]
  */
-BlockchainDouble.prototype.processBlock = function(block, commit, callback) {
+BlockchainDouble.prototype.processBlock = function(vm, block, commit, callback) {
   var self = this;
 
-  if (typeof commit == "function") {
+  if (typeof commit === "function") {
     callback = commit;
     commit = true;
   }
 
-  var currentStateRoot = self.stateTrie.root;
+  vm.runBlock(
+    {
+      block: block,
+      generate: true,
+      skipBlockValidation: true
+    },
+    async function(vmerr, results) {
+      // This is a check that has been in there for awhile. I'm unsure if it's required, but it can't hurt.
+      if (vmerr && vmerr instanceof Error === false) {
+        vmerr = new Error("VM error: " + vmerr);
+      }
 
-  // We checkpoint here for speed. We want all state trie reads/writes to happen in memory,
-  // and the final output be flushed to the database at the end of transaction processing.
-  self.vm.stateManager.checkpoint();
+      // If we're given an error back directly, it's worse than a runtime error. Expose it and get out.
+      if (vmerr) {
+        callback(vmerr);
+        return;
+      }
+      // If no error, check for a runtime error. This can return null if no runtime error.
+      vmerr = RuntimeError.fromResults(block.transactions, results);
 
-  var cleanup = function(err) {
-    self.vm.stateManager.revert(function(e) {
-      callback(err || e);
-    });
-  };
+      // Note, even if we have an error, some transactions may still have succeeded.
+      // Process their logs if so, returning the error at the end.
 
-  self.vm.runBlock({
-    block: block,
-    generate: true,
-  }, function(vmerr, results) {
-    // This is a check that has been in there for awhile. I'm unsure if it's required, but it can't hurt.
-    if (vmerr && vmerr instanceof Error == false) {
-      vmerr = new Error("VM error: " + vmerr);
-    }
+      var logs = [];
+      var receipts = [];
 
-    // If we're given an error back directly, it's worse than a runtime error. Expose it and get out.
-    if (vmerr) return cleanup(vmerr);
+      var totalBlockGasUsage = 0;
 
-    // If no error, check for a runtime error. This can return null if no runtime error.
-    vmerr = RuntimeError.fromResults(block.transactions, results);
+      results.results.forEach(function(result) {
+        totalBlockGasUsage += to.number(result.gasUsed);
+      });
 
-    // Note, even if we have an error, some transactions may still have succeeded.
-    // Process their logs if so, returning the error at the end.
+      block.header.gasUsed = utils.toBuffer(to.hex(totalBlockGasUsage));
 
-    var logs = [];
-    var receipts = [];
+      const txTrie = new Trie();
+      const rcptTrie = new Trie();
+      const promises = [];
+      const putInTrie = (trie, key, val) => promisify(trie.put.bind(trie))(key, val);
 
-    var totalBlockGasUsage = 0;
+      for (var v = 0; v < results.receipts.length; v++) {
+        var result = results.results[v];
+        var receipt = results.receipts[v];
+        var tx = block.transactions[v];
+        var txHash = tx.hash();
+        var txLogs = [];
 
-    results.results.forEach(function(result) {
-      totalBlockGasUsage += to.number(result.gasUsed);
-    });
+        // Only process the transaction's logs if it didn't error.
+        if (result.vm.exception === 1) {
+          for (var i = 0; i < receipt.logs.length; i++) {
+            var receiptLog = receipt.logs[i];
+            var address = to.hex(receiptLog[0]);
+            var topics = [];
 
-    block.header.gasUsed = utils.toBuffer(to.hex(totalBlockGasUsage));
+            for (var j = 0; j < receiptLog[1].length; j++) {
+              topics.push(to.hex(receiptLog[1][j]));
+            }
 
-    for (var v = 0; v < results.receipts.length; v++) {
-      var result = results.results[v];
-      var receipt = results.receipts[v];
-      var tx = block.transactions[v];
-      var tx_hash = tx.hash();
-      var tx_logs = [];
+            var data = to.hex(receiptLog[2]);
 
-      // Only process the transaction's logs if it didn't error.
-      if (result.vm.exception == 1) {
-        for (var i = 0; i < receipt.logs.length; i++) {
-          var log = receipt.logs[i];
-          var address = to.hex(log[0]);
-          var topics = []
+            var log = new Log({
+              logIndex: to.hex(i),
+              transactionIndex: to.hex(v),
+              transactionHash: txHash,
+              block: block,
+              address: address,
+              data: data,
+              topics: topics,
+              type: "mined"
+            });
 
-          for (var j = 0; j < log[1].length; j++) {
-            topics.push(to.hex(log[1][j]));
+            logs.push(log);
+            txLogs.push(log);
           }
-
-          var data = to.hex(log[2]);
-
-          var log = new Log({
-            logIndex: to.hex(i),
-            transactionIndex: to.hex(v),
-            transactionHash: tx_hash,
-            block: block,
-            address: address,
-            data: data,
-            topics: topics,
-            type: "mined"
-          });
-
-          logs.push(log);
-          tx_logs.push(log);
         }
+
+        let rcpt = new Receipt(
+          tx,
+          block,
+          txLogs,
+          result.gasUsed.toArrayLike(Buffer),
+          receipt.gasUsed,
+          result.createdAddress,
+          receipt.status,
+          to.hex(receipt.bitvector)
+        );
+        receipts.push(rcpt);
+
+        const rawReceipt = [receipt.status, receipt.gasUsed, receipt.bitvector, receipt.logs];
+        const rcptBuffer = utils.rlp.encode(rawReceipt);
+        const key = utils.rlp.encode(v);
+        promises.push(putInTrie(txTrie, key, tx.serialize()));
+        promises.push(putInTrie(rcptTrie, key, rcptBuffer));
       }
+      await Promise.all(promises);
 
-      receipts.push(new Receipt(tx, block, tx_logs, receipt.gasUsed, result.createdAddress, receipt.status));
-    }
+      block.header.transactionsTrie = utils.toBuffer(txTrie.root);
+      block.header.receiptTrie = utils.toBuffer(rcptTrie.root);
 
-    function commmitIfNeeded(cb) {
-      if (commit === true) {
-        self.vm.stateManager.commit(function(e) {
-          if (e) return cleanup(e);
-
-          // Put that block on the end the chain
-          self.putBlock(block, logs, receipts, cb);
-        });
+      if (commit) {
+        // Put that block on the end of the chain
+        self.putBlock(block, logs, receipts, done);
       } else {
-        self.vm.stateManager.revert(cb);
+        done();
+      }
+
+      function done(e) {
+        if (e) {
+          return callback(e);
+        }
+        // Note we return the vm err here too, if it exists.
+        callback(vmerr, block.transactions, results);
       }
     }
-
-    commmitIfNeeded(function(e) {
-      if (e) return callback(e);
-      // Note we return the vm err here too, if it exists.
-      callback(vmerr, block.transactions, results);
-    });
-  });
+  );
 };
 
 /**
@@ -575,15 +705,19 @@ BlockchainDouble.prototype.processBlock = function(block, commit, callback) {
  * Process the next block like a normal blockchain, pulling from the list of
  * pending transactions.
  *
+ * @param  {number} timestamp at which the block is mined
  * @param  {Function} callback Callback when transaction processing is finished.
  * @return [type]              [description]
  */
-BlockchainDouble.prototype.processNextBlock = function(callback) {
+BlockchainDouble.prototype.processNextBlock = function(timestamp, callback) {
   var self = this;
 
-  self.sortByPriceAndNonce();
+  if (typeof timestamp === "function") {
+    callback = timestamp;
+    timestamp = undefined;
+  }
 
-  var successfullyAddedTransactions = [];
+  self.sortByPriceAndNonce();
 
   // Grab only the transactions that can fit within the block
   var currentTransactions = [];
@@ -607,20 +741,39 @@ BlockchainDouble.prototype.processNextBlock = function(callback) {
   // Remember, we ensured transactions had a valid gas limit when they were queued (in the state manager).
   // If we run into a case where we can't process any because one is higher than the gas limit,
   // then it's a serious issue. This should never happen, but let's check anyway.
-  if (currentTransactions.length == 0 && self.pending_transactions.length > 0) {
+  if (currentTransactions.length === 0 && self.pending_transactions.length > 0) {
     // Error like geth.
-    return callback("Unexpected error condition: next transaction exceeds block gas limit")
+    var error = "Unexpected error condition: next transaction exceeds block gas limit";
+    return callback(error);
   }
 
   // Create a new block meant to be the end of the chain
   this.createBlock(function(err, block) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
+    // Overwrite block timestamp
+    if (timestamp) {
+      self.data.blocks.last(function(err, last) {
+        if (err) {
+          // it is safe to ignore this error as we only use the result
+          // to log a warning to the console.
+          return;
+        }
+        if (last && to.number(last.header.timestamp) > timestamp) {
+          self.logger.log(
+            "Waring: Setting the block timestamp (" + timestamp + ") that is earlier than the parent block one."
+          );
+        }
+      });
+      block.header.timestamp = to.hex(timestamp);
+      self.setTime(new Date(timestamp * 1000));
+    }
     // Add transactions to the block.
     Array.prototype.push.apply(block.transactions, currentTransactions);
 
-    // Process the block, committing the block to the chain
-    self.processBlock(block, true, callback);
+    self.processBlock(self.vm, block, true, callback);
   });
 };
 
@@ -635,135 +788,179 @@ BlockchainDouble.prototype.processNextBlock = function(callback) {
  *  1. Find block where transaction occurred
  *  2. Set state root of that block
  *  3. Rerun every transaction in that block prior to and including the requested transaction
- *  4. Reset state root back to original
- *  5. Send trace results back.
+ *  4. Send trace results back.
  *
  * @param  {[type]}   tx       [description]
  * @param  {Function} callback [description]
  * @return [type]              [description]
  */
-BlockchainDouble.prototype.processTransactionTrace = function(hash, params, callback) {
-  var self = this;
-  var target_hash = to.hex(hash);
-  var tx_hash_currently_processing = "";
-  var tx_currently_processing = null;
+BlockchainDouble.prototype.processTransactionTrace = async function(hash, params, callback) {
+  const self = this;
+  const targetHash = to.hex(hash);
+  let txHashCurrentlyProcessing = "";
+  let txCurrentlyProcessing = null;
+  let vm;
 
-  var returnVal = {
+  let storageStack = {
+    currentDepth: -1,
+    stack: []
+  };
+
+  let returnVal = {
     gas: 0,
     returnValue: "",
     structLogs: []
   };
 
-  function step_listener(event) {
+  function stepListener(event, next) {
     // See these docs:
     // https://github.com/ethereum/go-ethereum/wiki/Management-APIs
 
-    var gasLeft = to.number(event.gasLeft);
-    var totalGasUsedAfterThisStep = to.number(tx_currently_processing.gasLimit) - gasLeft;
-    var gasUsedThisStep = totalGasUsedAfterThisStep - returnVal.gas;
+    const gasLeft = to.number(event.gasLeft);
+    const totalGasUsedAfterThisStep = to.number(txCurrentlyProcessing.gasLimit) - gasLeft;
+    const gasUsedThisStep = totalGasUsedAfterThisStep - returnVal.gas;
     returnVal.gas += gasUsedThisStep;
 
-    // Get memory and break it up into 32-byte words.
-    // Note we may possibly have to pad the final word.
-    var memory = (new Buffer(event.memory)).toString("hex");
-    memory = memory.match(/.{1,64}/g) || [];
+    let memory = null;
+    if (!params.disableMemory) {
+      // Get memory and break it up into 32-byte words.
+      // Note we may possibly have to pad the final word.
+      memory = Buffer.from(event.memory).toString("hex");
+      memory = memory.match(/.{1,64}/g) || [];
 
-    if (memory.length > 0) {
-      var lastItem = memory[memory.length - 1];
-      if (lastItem.length < 64) {
-        memory[memory.length - 1] = lastItem + new Array(64 - lastItem.length + 1).join("0");
+      if (memory.length > 0) {
+        const lastItem = memory[memory.length - 1];
+        if (lastItem.length < 64) {
+          memory[memory.length - 1] = lastItem + new Array(64 - lastItem.length + 1).join("0");
+        }
       }
     }
 
-    var structLog = {
+    let stack = null;
+    if (!params.disableStack) {
+      stack = event.stack.map((item) => {
+        return to.rpcDataHexString(item, 64).replace("0x", ""); // non-0x prefixed.
+      });
+    }
+
+    let structLog = {
       depth: event.depth,
       error: "",
       gas: gasLeft,
       gasCost: gasUsedThisStep,
-      memory: memory,
+      memory,
       op: event.opcode.name,
       pc: event.pc,
-      stack: event.stack.map(function(item) {
-        return item.toString("hex"); // non-0x prefixed.
-      }),
-      storage: {} // TODO: Figure out storage
-    }
+      stack,
+      storage: null
+    };
 
-    returnVal.structLogs.push(structLog);
-  }
-
-  function beforeTx_listener(tx) {
-    tx_currently_processing = tx;
-    tx_hash_currently_processing = to.hex(tx.hash());
-
-    if (tx_hash_currently_processing == target_hash) {
-      self.vm.on("step", step_listener);
-    }
-  }
-
-  // afterTx_listener cleans up everything.
-  function afterTx_listener() {
-    if (tx_hash_currently_processing == target_hash) {
-      self.vm.removeListener("step", step_listener);
-      self.vm.removeListener("beforeTx", beforeTx_listener);
-      self.vm.removeListener("afterTx", afterTx_listener);
+    if (params.disableStorage) {
+      returnVal.structLogs.push(structLog);
+      next();
+    } else {
+      structLog = self.processStorageTrace(structLog, storageStack, event, vm, function(err, structLog) {
+        if (err) {
+          return next(err);
+        }
+        returnVal.structLogs.push(structLog);
+        next();
+      });
     }
   }
-
-  // Listen to beforeTx and afterTx so we know when our target transaction
-  // is processing. These events will add the vent listener for getting the trace data.
-  self.vm.on("beforeTx", beforeTx_listener);
-  self.vm.on("afterTx", afterTx_listener);
 
   // #1 - get block via transaction receipt
-  this.getTransactionReceipt(target_hash, function(err, receipt) {
-    if (err) return callback(err);
-
-    if (!receipt) {
-      return callback(new Error("Unknown transaction " + target_hash));
+  this.getTransactionReceipt(targetHash, function(err, receipt) {
+    if (err) {
+      return callback(err);
     }
 
-    var targetBlock = receipt.block;
+    if (!receipt) {
+      return callback(new Error("Unknown transaction " + targetHash));
+    }
+
+    let targetBlock = receipt.block;
 
     // Get the parent of the target block
     self.getBlock(targetBlock.header.parentHash, function(err, parent) {
-      if (err) return callback(err);
-
-      var startingStateRoot = self.stateTrie.root;
+      if (err) {
+        return callback(err);
+      }
 
       // #2 - Set state root of original block
-      self.stateTrie.root = parent.header.stateRoot;
+      var stateTrie = self.createStateTrie(self.data.trie_db, parent.header.stateRoot, {
+        // when forking we need to make sure we also copy over the forkBlockNumber,
+        // otherwise some operations will request data from the main chain at the
+        // latest block
+        forkBlockNumber: to.number(parent.header.number)
+      });
+      vm = self.createVMFromStateTrie(stateTrie);
 
       // Prepare the "next" block with necessary transactions
-      self.createBlock(parent, function(err, block) {
-        if (err) return callback(err);
+      self.createBlock(parent, false, function(err, block) {
+        if (err) {
+          return callback(err);
+        }
 
         for (var i = 0; i < targetBlock.transactions.length; i++) {
           var tx = targetBlock.transactions[i];
-          block.transactions.push(tx)
+          block.transactions.push(tx);
 
           // After including the target transaction, that's all we need to do.
-          if (to.hex(tx.hash()) == target_hash) {
+          if (to.hex(tx.hash()) === targetHash) {
             break;
           }
         }
 
+        function beforeTxListener(tx) {
+          txCurrentlyProcessing = tx;
+          txHashCurrentlyProcessing = to.hex(tx.hash());
+          if (txHashCurrentlyProcessing === targetHash) {
+            vm.on("step", stepListener);
+          }
+        }
+
+        // afterTxListener cleans up everything.
+        function afterTxListener() {
+          if (txHashCurrentlyProcessing === targetHash) {
+            removeListeners();
+          }
+        }
+
+        function removeListeners() {
+          vm.removeListener("step", stepListener);
+          vm.removeListener("beforeTx", beforeTxListener);
+          vm.removeListener("afterTx", afterTxListener);
+        }
+
+        // Listen to beforeTx and afterTx so we know when our target transaction
+        // is processing. These events will add the event listener for getting the trace data.
+        vm.on("beforeTx", beforeTxListener);
+        vm.on("afterTx", afterTxListener);
+
+        // Don't even let the vm try to flush the block's _cache to the stateTrie.
+        // When forking some of the data that the traced function may request will
+        // exist only on the main chain. Because we pretty much lie to the VM by
+        // telling it we DO have data in our Trie, when we really don't, it gets
+        // lost during the commit phase when it traverses the "borrowed" data's
+        // trie (as it may not have a valid root). Because this is a trace, and we
+        // don't need to commit the data, duck punching the `flush` method (the
+        // simplest method I could find) is fine.
+        // Remove this and you may see the infamous
+        // `Uncaught TypeError: Cannot read property 'pop' of undefined` error!
+        vm.stateManager._cache.flush = (cb) => cb();
+
         // #3 - Process the block without committing the data.
-        self.processBlock(block, false, function(err, transactions, results) {
+        self.processBlock(vm, block, false, function(err) {
           // Ignore runtime errors, or else erroneous transactions can't be traced.
-          if (err && err.message.indexOf("VM Exception") == 0) {
+          if (err && err.message.indexOf("VM Exception") === 0) {
             err = null;
           }
 
-          // #4 - reset the state root.
-          self.stateTrie.root = startingStateRoot;
-
           // Just to be safe
-          self.vm.removeListener("beforeTx", beforeTx_listener);
-          self.vm.removeListener("afterTx", afterTx_listener);
-          self.vm.removeListener("step", step_listener);
+          removeListeners();
 
-          // #5 - send state results back
+          // #4 - send state results back
           callback(err, returnVal);
         });
       });
@@ -771,11 +968,69 @@ BlockchainDouble.prototype.processTransactionTrace = function(hash, params, call
   });
 };
 
+BlockchainDouble.prototype.processStorageTrace = function(structLog, storageStack, event, vm, callback) {
+  var name = event.opcode.name;
+
+  if (storageStack.currentDepth > event.depth) {
+    storageStack.stack.pop();
+  }
+  if (storageStack.currentDepth < event.depth) {
+    storageStack.stack.push({});
+  }
+
+  storageStack.currentDepth = event.depth;
+
+  var key;
+  var value;
+  switch (name) {
+    case "SSTORE":
+      {
+        const stack = event.stack;
+        const stackLength = stack.length;
+        key = to.rpcDataHexString(stack[stackLength - 1], 64).replace("0x", "");
+        value = to.rpcDataHexString(stack[stackLength - 2], 64).replace("0x", "");
+        // use Object.assign to prevent future steps from overwriting this step's storage values
+        structLog.storage = Object.assign({}, storageStack.stack[storageStack.currentDepth]);
+
+        callback(null, structLog);
+        // assign after callback because this storage change actually takes
+        // effect _after_ this opcode executes
+        storageStack.stack[storageStack.currentDepth][key] = value;
+      }
+      break;
+    case "SLOAD":
+      {
+        const stack = event.stack;
+        // this one's more fun, we need to get the value the contract is loading from current storage
+        key = to.rpcDataHexString(stack[stack.length - 1], 64).replace("0x", "");
+
+        vm.stateManager.getContractStorage(event.address, "0x" + key, function(err, result) {
+          if (err) {
+            return callback(err);
+          }
+
+          value = to.rpcDataHexString(result, 64).replace("0x", "");
+          storageStack.stack[storageStack.currentDepth][key] = value;
+          // use Object.assign to prevent future steps from overwriting this step's storage values
+          structLog.storage = Object.assign({}, storageStack.stack[storageStack.currentDepth]);
+          callback(null, structLog);
+        });
+      }
+      break;
+    default:
+      // use Object.assign to prevent future steps from overwriting this step's storage values
+      structLog.storage = Object.assign({}, storageStack.stack[storageStack.currentDepth]);
+      callback(null, structLog);
+  }
+};
+
 BlockchainDouble.prototype.getAccount = function(address, number, callback) {
   var self = this;
 
   this.getBlock(number, function(err, block) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     var trie = self.stateTrie;
 
@@ -787,11 +1042,11 @@ BlockchainDouble.prototype.getAccount = function(address, number, callback) {
       // Finally, put the stateRoot back for good
       trie.root = currentStateRoot;
 
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       var account = new Account(data);
-
-      account.exists = !!data;
 
       callback(null, account);
     });
@@ -800,14 +1055,18 @@ BlockchainDouble.prototype.getAccount = function(address, number, callback) {
 
 BlockchainDouble.prototype.getNonce = function(address, number, callback) {
   this.getAccount(address, number, function(err, account) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     callback(null, account.nonce);
   });
 };
 
 BlockchainDouble.prototype.getBalance = function(address, number, callback) {
   this.getAccount(address, number, function(err, account) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     callback(null, account.balance);
   });
@@ -818,7 +1077,9 @@ BlockchainDouble.prototype.getStorage = function(address, position, number, call
   var self = this;
 
   this.getBlock(number, function(err, block) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     var trie = self.stateTrie;
 
@@ -847,16 +1108,17 @@ BlockchainDouble.prototype.getStorage = function(address, position, number, call
 
         callback(null, value);
       });
-
     });
   });
-}
+};
 
 BlockchainDouble.prototype.getCode = function(address, number, callback) {
   var self = this;
 
   this.getBlock(number, function(err, block) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     var trie = self.stateTrie;
 
@@ -877,7 +1139,9 @@ BlockchainDouble.prototype.getCode = function(address, number, callback) {
         // Finally, put the stateRoot back for good
         trie.root = currentStateRoot;
 
-        if (err) return callback(err);
+        if (err) {
+          return callback(err);
+        }
 
         callback(null, code);
       });
@@ -902,15 +1166,21 @@ BlockchainDouble.prototype.getTransaction = function(hash, callback) {
 
 BlockchainDouble.prototype.getTransactionReceipt = function(hash, callback) {
   hash = to.hex(hash);
+  const pendingTxs = this.pending_transactions;
+
+  for (var i = 0; i < pendingTxs.length; i++) {
+    let pendingTxHash = to.hex(pendingTxs[i].hash());
+    if (hash === pendingTxHash) {
+      return callback(null, { tx: pendingTxs[i] });
+    }
+  }
 
   this.data.transactionReceipts.get(hash, function(err, receipt) {
     if (err) {
-      if (err) {
-        if (err.notFound) {
-          return callback(null, null);
-        } else {
-          return callback(err);
-        }
+      if (err.notFound) {
+        return callback(null, null);
+      } else {
+        return callback(err);
       }
     }
 
@@ -921,31 +1191,37 @@ BlockchainDouble.prototype.getTransactionReceipt = function(hash, callback) {
 BlockchainDouble.prototype.getBlockLogs = function(number, callback) {
   var self = this;
   this.getEffectiveBlockNumber(number, function(err, effective) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     self.data.blockLogs.get(effective, callback);
   });
 };
 
 BlockchainDouble.prototype.getHeight = function(callback) {
   this.data.blocks.length(function(err, length) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     callback(null, length - 1);
-  })
+  });
 };
 
 BlockchainDouble.prototype.currentTime = function() {
-  return (new Date().getTime() / 1000 | 0) + this.timeAdjustment;
+  return ((new Date().getTime() / 1000) | 0) + this.timeAdjustment;
 };
 
 BlockchainDouble.prototype.increaseTime = function(seconds) {
-  if (seconds < 0) seconds = 0;
+  if (seconds < 0) {
+    seconds = 0;
+  }
   this.timeAdjustment += seconds;
   return this.timeAdjustment;
 };
 
 BlockchainDouble.prototype.setTime = function(date) {
-  var now = new Date().getTime() / 1000 | 0;
-  var start = date.getTime() / 1000 | 0;
+  var now = (new Date().getTime() / 1000) | 0;
+  var start = (date.getTime() / 1000) | 0;
   this.timeAdjustment = start - now;
 };
 
