@@ -8,9 +8,21 @@ module.exports = {
     var buf = [];
 
     for (var i = 0; i < length; i++) {
-      buf.push(rng()*255);
+      buf.push(rng() * 255);
     }
 
-    return new Buffer(buf);
+    return Buffer.from(buf);
+  },
+
+  randomAlphaNumericString: function(length, rng) {
+    const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+    let text = "";
+
+    for (var i = 0; i < length; i++) {
+      text += alphabet.charAt(Math.floor((rng || Math.random)() * alphabet.length));
+    }
+
+    return text;
   }
-}
+};
