@@ -1,23 +1,20 @@
 var LevelUpArrayAdapter = require("./database/leveluparrayadapter");
 var LevelUpObjectAdapter = require("./database/levelupobjectadapter");
-var levelup = require('levelup');
+var levelup = require("levelup");
+var encode = require("encoding-down");
 var filedown = require("./database/filedown");
 var cachedown = require("cachedown");
-var Sublevel = require("level-sublevel");
-var Block = require("ethereumjs-block");
 var txserializer = require("./database/txserializer");
 var blockserializer = require("./database/blockserializer");
 var bufferserializer = require("./database/bufferserializer");
 var BlockLogsSerializer = require("./database/blocklogsserializer");
 var ReceiptSerializer = require("./database/receiptserializer");
-var to = require("./utils/to");
-var utils = require("ethereumjs-util");
-var FakeTransaction = require('ethereumjs-tx/fake.js');
 var tmp = require("tmp");
 
 function Database(options) {
   this.options = options;
-};
+  this.directory = null;
+}
 
 Database.prototype.initialize = function(callback) {
   var self = this;
@@ -31,22 +28,24 @@ Database.prototype.initialize = function(callback) {
   }
 
   getDir(function(err, directory) {
-    if (err) return callback(err);
-
-    levelup(directory, {
-      valueEncoding: "json",
-      db: function (location) {
-        // This cache size was chosen from a plethora of hand testing.
-        // It seems a not-too-large cache (100) size is the right amount.
-        // When higher (say 10000), it seems the benefits wear off.
-        // See /perf/transactions.js for a benchmark.
-        return cachedown(location, filedown).maxSize(100);
-      }
-    }, finishInitializing)
+    if (err) {
+      return callback(err);
+    }
+    const levelupOptions = { valueEncoding: "json" };
+    if (self.options.db) {
+      const store = self.options.db;
+      levelup(store, levelupOptions, finishInitializing);
+    } else {
+      self.directory = directory;
+      const store = encode(cachedown(directory, filedown).maxSize(100), levelupOptions);
+      levelup(store, {}, finishInitializing);
+    }
   });
 
   function finishInitializing(err, db) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     self.db = db;
 
@@ -68,7 +67,7 @@ Database.prototype.initialize = function(callback) {
     self.trie_db = new LevelUpObjectAdapter("trie_db", self.db, bufferserializer, bufferserializer);
 
     callback();
-  };
+  }
 };
 
 Database.prototype.close = function(callback) {
