@@ -7,6 +7,6 @@ module.exports = {
     cb(null, hex);
   },
   decode: function(json, cb) {
-    cb(null, utils.toBuffer(json))
+    cb(null, utils.toBuffer(json));
   }
 };
