@@ -7,11 +7,15 @@ function LevelUpValueAdapter(name, db, serializer) {
   this.db = this.db.sublevel(name);
   this.name = name;
   this.serializer = serializer || {
-    encode: function(val, callback) { callback(null, val); },
-    decode: function(val, callback) { callback(null, val); }
+    encode: function(val, callback) {
+      callback(null, val);
+    },
+    decode: function(val, callback) {
+      callback(null, val);
+    }
   };
   this.value_key = "value";
-};
+}
 
 LevelUpValueAdapter.prototype.get = function(callback) {
   var self = this;
@@ -32,13 +36,15 @@ LevelUpValueAdapter.prototype.get = function(callback) {
 LevelUpValueAdapter.prototype.set = function(value, callback) {
   var self = this;
   this.serializer.encode(value, function(err, encoded) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     self.db.put(self.value_key, encoded, callback);
   });
 };
 
 LevelUpValueAdapter.prototype.del = function(callback) {
   this.db.del(this.value_key, callback);
-}
+};
 
 module.exports = LevelUpValueAdapter;
