@@ -1,4 +1,4 @@
-var Subprovider = require('web3-provider-engine/subproviders/subprovider.js');
+var Subprovider = require("web3-provider-engine/subproviders/subprovider.js");
 var inherits = require("util").inherits;
 
 inherits(ReactiveBlockTracker, Subprovider);
@@ -7,11 +7,11 @@ module.exports = ReactiveBlockTracker;
 
 function ReactiveBlockTracker() {
   this.methods = {
-    "eth_call": "before",
-    "eth_getStorageAt": "before",
-    "eth_getLogs": "before"
+    eth_call: "before",
+    eth_getStorageAt: "before",
+    eth_getLogs: "before"
   };
-};
+}
 
 // Fetch the block before certain requests to make sure we're completely updated
 // before those methods are processed. Also, fetch the block after certain requests
@@ -26,21 +26,26 @@ ReactiveBlockTracker.prototype.handleRequest = function(payload, next, end) {
   }
 
   function fetchBlock(cb) {
-    self.engine._fetchBlock("latest", function(err, block) {
-      if (err) return end(err);
-      if (!self.engine.currentBlock || 0 !== self.engine.currentBlock.hash.compare(block.hash)) {
+    self.engine.fetchBlock("latest", function(err, block) {
+      if (err) {
+        return end(err);
+      }
+      if (!self.engine.currentBlock || self.engine.currentBlock.hash.compare(block.hash) !== 0) {
         self.engine._setCurrentBlock(block);
       }
       cb();
     });
   }
 
-  if (when == "before") {
+  if (when === "before") {
     fetchBlock(function() {
       next();
     });
   } else {
     next(function(error, result, cb) {
+      if (error) {
+        return cb(error);
+      }
       fetchBlock(cb);
     });
   }
