@@ -1,4 +1,4 @@
-var Subprovider = require('web3-provider-engine/subproviders/subprovider.js');
+var Subprovider = require("web3-provider-engine/subproviders/subprovider.js");
 var inherits = require("util").inherits;
 
 inherits(RequestFunnel, Subprovider);
@@ -11,30 +11,30 @@ module.exports = RequestFunnel;
 function RequestFunnel() {
   // We use an object here for O(1) lookups (speed).
   this.methods = {
-    "eth_call": true,
-    "eth_getStorageAt": true,
-    "eth_sendTransaction": true,
-    "eth_sendRawTransaction": true,
+    eth_call: true,
+    eth_getStorageAt: true,
+    eth_sendTransaction: true,
+    eth_sendRawTransaction: true,
 
     // Ensure block filter and filter changes are process one at a time
     // as well so filter requests that come in after a transaction get
     // processed once that transaction has finished processing.
-    "eth_newBlockFilter": true,
-    "eth_getFilterChanges": true,
-    "eth_getFilterLogs": true
+    eth_newBlockFilter: true,
+    eth_getFilterChanges: true,
+    eth_getFilterLogs: true
   };
   this.queue = [];
   this.isWorking = false;
-};
+}
 
 RequestFunnel.prototype.handleRequest = function(payload, next, end) {
-  if (payload.external != true || this.methods[payload.method] != true) {
+  if (payload.external !== true || this.methods[payload.method] !== true) {
     return next();
   }
 
   this.queue.push([payload, next]);
 
-  if (this.isWorking == false) {
+  if (this.isWorking === false) {
     this.processNext();
   }
 };
@@ -42,7 +42,7 @@ RequestFunnel.prototype.handleRequest = function(payload, next, end) {
 RequestFunnel.prototype.processNext = function() {
   var self = this;
 
-  if (this.queue.length == 0) {
+  if (this.queue.length === 0) {
     this.isWorking = false;
     return;
   }
@@ -50,10 +50,12 @@ RequestFunnel.prototype.processNext = function() {
   this.isWorking = true;
 
   var item = this.queue.shift();
-  var payload = item[0];
   var next = item[1];
 
   next(function(error, request, cb) {
+    if (error) {
+      return cb(error);
+    }
     cb();
     self.processNext();
   });
