@@ -1,39 +1,36 @@
 var BlockchainDouble = require("../blockchain_double.js");
-var VM = require("ethereumjs-vm");
-var Account = require("ethereumjs-account");
+var Account = require("ethereumjs-account").default;
 var Block = require("ethereumjs-block");
 var Log = require("./log.js");
 var Receipt = require("./receipt.js");
 var utils = require("ethereumjs-util");
 var ForkedStorageTrie = require("./forkedstoragetrie.js");
-var FakeTransaction = require('ethereumjs-tx/fake.js');
 var Web3 = require("web3");
 var to = require("./to.js");
+var Transaction = require("./transaction");
 var async = require("async");
-var txhelper = require("./txhelper.js")
+const BN = utils.BN;
 
 var inherits = require("util").inherits;
 
 inherits(ForkedBlockchain, BlockchainDouble);
 
 function ForkedBlockchain(options) {
-  var self = this;
-
-  options = options || {};
+  this.options = options || {};
 
   if (options.fork == null) {
     throw new Error("ForkedBlockchain must be passed a fork parameter.");
   }
 
   this.fork = options.fork;
-  this.fork_block_number = options.fork_block_number;
-  this.fork_version = null;
+  this.forkBlockNumber = options.fork_block_number;
+  this.forkVersion = null;
 
-  if (typeof this.fork == "string") {
+  if (typeof this.fork === "string") {
     if (this.fork.indexOf("@") >= 0) {
       var split = this.fork.split("@");
       this.fork = split[0];
-      this.fork_block_number = parseInt(split[1]);
+      this.forkBlockNumber = parseInt(split[1]);
     }
 
     this.fork = new Web3.providers.HttpProvider(this.fork);
@@ -44,62 +41,86 @@ function ForkedBlockchain(options) {
 
   BlockchainDouble.call(this, options);
 
+  this.createVMFromStateTrie = function() {
+    var vm = BlockchainDouble.prototype.createVMFromStateTrie.apply(this, arguments);
+    this.patchVM(vm);
+    return vm;
+  };
+
   this.web3 = new Web3(this.fork);
-};
+}
 
 ForkedBlockchain.prototype.initialize = function(accounts, callback) {
   var self = this;
 
-  this.web3.version.getNetwork(function(err, version) {
-    if (err) return callback(err);
+  this.web3.eth.net.getId(function(err, version) {
+    if (err) {
+      return callback(err);
+    }
 
-    self.fork_version = version;
+    self.forkVersion = version;
 
-    BlockchainDouble.prototype.initialize.call(self, accounts, function(err) {
-      if (err) return callback(err);
-
-      // Unfortunately forking requires a bit of monkey patching, but it gets the job done.
-      self.vm.stateManager._lookupStorageTrie = self.lookupStorageTrie.bind(self);
-      self.vm.stateManager.cache._lookupAccount = self.getAccount.bind(self);
-      self.vm.stateManager.getContractCode = self.getCode.bind(self);
-      self.vm.stateManager.putContractCode = self.putCode.bind(self);
-
-      callback();
-    });
+    BlockchainDouble.prototype.initialize.call(self, accounts, callback);
   });
 };
+ForkedBlockchain.prototype.patchVM = function(vm) {
+  const trie = vm.stateManager._trie;
+  const lookupAccount = this.getLookupAccount(trie);
+  // Unfortunately forking requires a bit of monkey patching, but it gets the job done.
+  vm.stateManager._cache._lookupAccount = lookupAccount;
+  vm.stateManager._lookupStorageTrie = this.getLookupStorageTrie(trie, lookupAccount);
+  vm.stateManager.getContractCode = this.getCode.bind(this);
+};
 
-ForkedBlockchain.prototype.createStateTrie = function(db, root) {
-  return new ForkedStorageTrie(db, root, {
-    fork: this.fork,
-    fork_block_number: this.fork_block_number,
-    blockchain: this
-  });
+/**
+ * @param db
+ * @param root
+ * @param options Allows overriding the options passed to the ForkedStorageTrie,
+ * like `forkBlockNumber` (required for tracing transactions)
+ */
+ForkedBlockchain.prototype.createStateTrie = function(db, root, options) {
+  options = Object.assign(
+    {
+      fork: this.fork,
+      forkBlockNumber: this.forkBlockNumber,
+      blockchain: this
+    },
+    options
+  );
+  // never allow the forkBlockNumber to go beyond our root forkBlockNumber
+  if (options.forkBlockNumber > this.forkBlockNumber) {
+    options.forkBlockNumber = this.forkBlockNumber;
+  }
+  return new ForkedStorageTrie(db, root, options);
 };
 
 ForkedBlockchain.prototype.createGenesisBlock = function(callback) {
   var self = this;
-  var blockNumber = this.fork_block_number || "latest";
+  var blockNumber = this.forkBlockNumber || "latest";
 
   self.web3.eth.getBlock(blockNumber, function(err, json) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     // If no start time was passed, set the time to where we forked from.
     // We only want to do this if a block was explicitly passed. If a block
     // number wasn't passed, then we're using the last block and the current time.
-    if (!self.time && self.fork_block_number) {
-      self.time = new Date(to.number(json.timestamp) * 1000);
+    if (!self.time && self.forkBlockNumber) {
+      self.time = self.options.time = new Date(to.number(json.timestamp) * 1000);
       self.setTime(self.time);
     }
 
-    blockNumber = to.hex(json.number);
+    blockNumber = json.number;
 
     // Update the relevant block numbers
-    self.fork_block_number = blockNumber;
-    self.stateTrie.fork_block_number = blockNumber;
+    self.forkBlockNumber = self.options.fork_block_number = blockNumber;
+    self.stateTrie.forkBlockNumber = self.stateTrie.options.forkBlockNumber = blockNumber;
 
     self.createBlock(function(err, block) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       block.header.number = utils.toBuffer(to.number(json.number) + 1);
       block.header.parentHash = utils.toBuffer(json.hash);
@@ -109,53 +130,48 @@ ForkedBlockchain.prototype.createGenesisBlock = function(callback) {
   });
 };
 
-ForkedBlockchain.prototype.createForkedStorageTrie = function(address) {
-  address = to.hex(address);
+ForkedBlockchain.prototype.getLookupStorageTrie = function(stateTrie, lookupAccount) {
+  lookupAccount = lookupAccount || this.getLookupAccount(stateTrie);
+  return (address, callback) => {
+    const storageTrie = stateTrie.copy();
+    storageTrie.address = address;
+    lookupAccount(address, (err, account) => {
+      if (err) {
+        return callback(err);
+      }
 
-  var trie = new ForkedStorageTrie(this.data.trie_db, null, {
-    address: address,
-    stateTrie: this.stateTrie,
-    blockchain: this,
-    fork: this.fork,
-    fork_block_number: this.fork_block_number
-  });
-
-  this.storageTrieCache[address] = trie;
-
-  return trie;
-};
-
-ForkedBlockchain.prototype.lookupStorageTrie = function(address, callback) {
-  var self = this
-
-  address = to.hex(address);
-
-  if (this.storageTrieCache[address] != null) {
-    return callback(null, this.storageTrieCache[address]);
-  }
-
-  callback(null, this.createForkedStorageTrie(address));
+      storageTrie.root = account.stateRoot;
+      callback(null, storageTrie);
+    });
+  };
 };
 
 ForkedBlockchain.prototype.isFallbackBlock = function(value, callback) {
   var self = this;
 
   self.getEffectiveBlockNumber(value, function(err, number) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
-    callback(null, number <= to.number(self.fork_block_number));
+    callback(null, number <= to.number(self.forkBlockNumber));
   });
 };
 
 ForkedBlockchain.prototype.isBlockHash = function(value) {
-  return typeof value == "string" && value.indexOf("0x") == 0 && value.length > 42;
-}
+  const isHash = typeof value === "string" && value.indexOf("0x") === 0 && value.length > 42;
+  return isHash || (Buffer.isBuffer(value) && value.byteLength > 20);
+};
 
 ForkedBlockchain.prototype.isFallbackBlockHash = function(value, callback) {
   var self = this;
 
   if (!this.isBlockHash(value)) {
     return callback(null, false);
+  }
+
+  if (Buffer.isBuffer(value)) {
+    value = to.hex(value);
   }
 
   self.data.blockHashes.get(value, function(err, blockIndex) {
@@ -169,22 +185,25 @@ ForkedBlockchain.prototype.isFallbackBlockHash = function(value, callback) {
     }
     callback(null, false);
   });
-}
+};
 
-ForkedBlockchain.prototype.getFallbackBlock = function(number_or_hash, cb) {
+ForkedBlockchain.prototype.getFallbackBlock = function(numberOrHash, cb) {
   var self = this;
 
-  // This function sometimes gets passed hex values that aren't long enough to
-  // be a block hash. In that case, let's convert them to a big number so Web3
-  // doesn't get stopped up.
-  if (typeof number_or_hash == "string" && number_or_hash.length < 66) {
-    number_or_hash = self.web3.toBigNumber(number_or_hash)
+  if (Buffer.isBuffer(numberOrHash)) {
+    // When tracing a transaction the VM sometimes ask for a block numbers as
+    // buffers instead of numbers.
+    numberOrHash = to.rpcDataHexString(numberOrHash);
   }
 
-  self.web3.eth.getBlock(number_or_hash, true, function(err, json) {
-    if (err) return cb(err);
+  self.web3.eth.getBlock(numberOrHash, true, function(err, json) {
+    if (err) {
+      return cb(err);
+    }
 
-    if (json == null) return cb();
+    if (json == null) {
+      return cb();
+    }
 
     var block = new Block();
 
@@ -192,7 +211,7 @@ ForkedBlockchain.prototype.getFallbackBlock = function(number_or_hash, cb) {
     block.header.uncleHash = utils.toBuffer(json.sha3Uncles);
     block.header.coinbase = utils.toBuffer(json.miner);
     block.header.stateRoot = utils.toBuffer(json.stateRoot); // Should we include the following three?
-    block.header.transactionTrie = utils.toBuffer(json.transactionsRoot);
+    block.header.transactionsTrie = utils.toBuffer(json.transactionsRoot);
     block.header.receiptTrie = utils.toBuffer(json.receiptsRoot);
     block.header.bloom = utils.toBuffer(json.logsBloom);
     block.header.difficulty = utils.toBuffer("0x" + json.totalDifficulty.toString(16)); // BigNumber
@@ -202,67 +221,68 @@ ForkedBlockchain.prototype.getFallbackBlock = function(number_or_hash, cb) {
     block.header.timestamp = utils.toBuffer(json.timestamp);
     block.header.extraData = utils.toBuffer(json.extraData);
 
-    (json.transactions || []).forEach(function(tx_json, index) {
-      block.transactions.push(txhelper.fromJSON(tx_json));
+    (json.transactions || []).forEach(function(txJson, index) {
+      block.transactions.push(Transaction.fromJSON(txJson, Transaction.types.real));
     });
 
     // Fake block. Let's do the worst.
     // TODO: Attempt to fill out all block data so as to produce the same hash! (can we?)
     block.hash = function() {
       return utils.toBuffer(json.hash);
-    }
+    };
 
     cb(null, block);
   });
-}
-
+};
 
 ForkedBlockchain.prototype.getBlock = function(number, callback) {
-  var self = this;
-
-  this.isFallbackBlockHash(number, function(err, isFallbackBlockHash) {
-    if (err) return callback(err);
-    if (isFallbackBlockHash) {
-      return self.getFallbackBlock(number, callback);
+  let checkFn;
+  const isBlockHash = this.isBlockHash(number);
+  if (isBlockHash) {
+    checkFn = this.isFallbackBlockHash;
+  } else {
+    checkFn = this.isFallbackBlock;
+  }
+  checkFn.call(this, number, (err, isFallback) => {
+    if (err) {
+      return callback(err);
+    }
+    if (isFallback) {
+      return this.getFallbackBlock(number, callback);
     }
 
-    self.isFallbackBlock(number, function(err, isFallbackBlock) {
-      if (err) return callback(err);
-
-      if (isFallbackBlock) {
-        return self.getFallbackBlock(number, callback);
-      }
-
-      // If we don't have string-based a block hash, turn what we do have into a number
-      // before sending it to getBlock.
-      function getBlockReference(value, callback) {
-        if (!self.isBlockHash(value)) {
-          self.getRelativeBlockNumber(value, callback);
-        } else {
-          callback(null, value);
+    const getBlock = BlockchainDouble.prototype.getBlock.bind(this);
+    if (isBlockHash) {
+      getBlock(number, callback);
+    } else {
+      this.getRelativeBlockNumber(number, (err, number) => {
+        if (err) {
+          return callback(err);
         }
-      }
-
-      getBlockReference(number, function(err, blockReference) {
-        if (err) return callback(err);
-
-        BlockchainDouble.prototype.getBlock.call(self, blockReference, callback);
+        getBlock(number, callback);
       });
-    });
+    }
   });
 };
 
 ForkedBlockchain.prototype.getStorage = function(address, key, number, callback) {
-  this.lookupStorageTrie(address, function(err, trie) {
-    if (err) return callback(err);
-    trie.get(key, callback);
+  this.getLookupStorageTrie(this.stateTrie)(address, (err, trie) => {
+    if (err) {
+      return callback(err);
+    }
+    this.getEffectiveBlockNumber(number, (err, number) => {
+      if (err) {
+        return callback(err);
+      }
+      trie.get(utils.setLengthLeft(utils.toBuffer(key), 32), number, callback);
+    });
   });
 };
 
 ForkedBlockchain.prototype.getCode = function(address, number, callback) {
   var self = this;
 
-  if (typeof number == "function") {
+  if (typeof number === "function") {
     callback = number;
     number = "latest";
   }
@@ -272,18 +292,23 @@ ForkedBlockchain.prototype.getCode = function(address, number, callback) {
   }
 
   this.getEffectiveBlockNumber(number, function(err, effective) {
+    if (err) {
+      return callback(err);
+    }
     number = effective;
 
     self.stateTrie.keyExists(address, function(err, exists) {
+      if (err) {
+        return callback(err);
+      }
       // If we've stored the value and we're looking at one of our stored blocks,
       // get it from our stored data.
-      if (exists && number > to.number(self.fork_block_number)) {
+      if (exists && number > to.number(self.forkBlockNumber)) {
         BlockchainDouble.prototype.getCode.call(self, address, number, callback);
       } else {
-
         // Else, we need to fetch it from web3. If our number is greater than
         // the fork, let's just use "latest".
-        if (number > to.number(self.fork_block_number)) {
+        if (number > to.number(self.forkBlockNumber)) {
           number = "latest";
         }
 
@@ -298,51 +323,53 @@ ForkedBlockchain.prototype.getCode = function(address, number, callback) {
   });
 };
 
-ForkedBlockchain.prototype.putCode = function(address, value, callback) {
-  // This is a bit of a hack. We need to bypass the vm's
-  // _lookupAccount call that vm.stateManager.putContractCode() uses.
-  // This means we have to do some things ourself. The last call
-  // to self.stateTrie.put() at the bottom is important because
-  // we can't just be satisfied putting it in the cache.
-
-  var self = this;
-  address = utils.toBuffer(address);
-  this.stateTrie.get(address, function(err, data) {
-    if (err) return callback(err);
-
-    var account = new Account(data);
-    account.setCode(self.stateTrie, value, function(err, result) {
-      if (err) return callback(err);
-
-      self.stateTrie.put(address, account.serialize(), function(err) {
-        if (err) return callback(err);
-
-        // Ensure the cache updates as well.
-        self.vm.stateManager._putAccount(address, account, callback);
-      });
+ForkedBlockchain.prototype.getLookupAccount = function(trie) {
+  return (address, callback) => {
+    // If the account doesn't exist in our state trie, get it off the wire.
+    trie.keyExists(address, (err, exists) => {
+      if (err) {
+        return callback(err);
+      }
+      if (exists) {
+        trie.get(address, (err, data) => {
+          if (err) {
+            return callback(err);
+          }
+          const account = new Account(data);
+          callback(null, account);
+        });
+      } else {
+        this.fetchAccountFromFallback(address, to.number(trie.forkBlockNumber), callback);
+      }
     });
-  })
+  };
 };
 
 ForkedBlockchain.prototype.getAccount = function(address, number, callback) {
   var self = this;
 
-  if (typeof number == "function") {
+  if (typeof number === "function") {
     callback = number;
     number = "latest";
   }
 
   this.getEffectiveBlockNumber(number, function(err, effective) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     number = effective;
 
     // If the account doesn't exist in our state trie, get it off the wire.
     self.stateTrie.keyExists(address, function(err, exists) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
-      if (exists && number > to.number(self.fork_block_number)) {
+      if (exists && number > to.number(self.forkBlockNumber)) {
         BlockchainDouble.prototype.getAccount.call(self, address, number, function(err, acc) {
-          if (err) return callback(err);
+          if (err) {
+            return callback(err);
+          }
           callback(null, acc);
         });
       } else {
@@ -355,14 +382,20 @@ ForkedBlockchain.prototype.getAccount = function(address, number, callback) {
 ForkedBlockchain.prototype.getTransaction = function(hash, callback) {
   var self = this;
   BlockchainDouble.prototype.getTransaction.call(this, hash, function(err, tx) {
-    if (err) return callback(err);
-    if (tx != null) return callback(null, tx);
+    if (err) {
+      return callback(err);
+    }
+    if (tx != null) {
+      return callback(null, tx);
+    }
 
     self.web3.eth.getTransaction(hash, function(err, result) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       if (result) {
-        result = txhelper.fromJSON(result);
+        result = Transaction.fromJSON(result, Transaction.types.signed);
       }
 
       callback(null, result);
@@ -373,132 +406,172 @@ ForkedBlockchain.prototype.getTransaction = function(hash, callback) {
 ForkedBlockchain.prototype.getTransactionReceipt = function(hash, callback) {
   var self = this;
   BlockchainDouble.prototype.getTransactionReceipt.call(this, hash, function(err, receipt) {
-    if (err) return callback(err);
-    if (receipt) return callback(null, receipt);
+    if (err) {
+      return callback(err);
+    }
+    if (receipt) {
+      return callback(null, receipt);
+    }
 
-    self.web3.eth.getTransactionReceipt(hash, function(err, receipt_json) {
-      if (err) return callback(err);
-      if (!receipt_json) return callback();
+    self.web3.eth.getTransactionReceipt(hash, function(err, receiptJson) {
+      if (err) {
+        return callback(err);
+      }
+      if (!receiptJson) {
+        return callback();
+      }
 
-      async.parallel({
-        tx: self.getTransaction.bind(self, hash),
-        block: self.getBlock.bind(self, receipt_json.blockNumber)
-      }, function(err, result) {
-        if (err) return callback(err);
+      async.parallel(
+        {
+          tx: self.getTransaction.bind(self, hash),
+          block: self.getBlock.bind(self, receiptJson.blockNumber)
+        },
+        function(err, result) {
+          if (err) {
+            return callback(err);
+          }
 
-        var logs = receipt_json.logs.map(function(log) {
-          return new Log(log);
-        });
+          var logs = receiptJson.logs.map(function(log) {
+            return new Log(log);
+          });
 
-        var receipt = new Receipt(result.tx, result.block, logs, receipt_json.cumulativeGasUsed, receipt_json.contractAddress, receipt_json.status);
+          var receipt = new Receipt(
+            result.tx,
+            result.block,
+            logs,
+            receiptJson.gasUsed,
+            receiptJson.cumulativeGasUsed,
+            receiptJson.contractAddress,
+            receiptJson.status,
+            to.hex(receiptJson.logsBloom)
+          );
 
-        callback(null, receipt);
+          callback(null, receipt);
+        }
+      );
+    });
+  });
+};
+
+ForkedBlockchain.prototype.fetchAccountFromFallback = function(address, blockNumber, callback) {
+  var self = this;
+  address = to.hex(address);
+
+  async.parallel(
+    {
+      code: this.fetchCodeFromFallback.bind(this, address, blockNumber),
+      balance: this.fetchBalanceFromFallback.bind(this, address, blockNumber),
+      nonce: this.fetchNonceFromFallback.bind(this, address, blockNumber)
+    },
+    function(err, results) {
+      if (err) {
+        return callback(err);
+      }
+
+      var code = results.code;
+      var balance = results.balance;
+      var nonce = results.nonce;
+
+      var account = new Account({
+        nonce: nonce,
+        balance: balance
       });
-    });
-  });
+
+      // This puts the code on the trie, keyed by the hash of the code.
+      // It does not actually link an account to code in the trie.
+      account.setCode(self.stateTrie, utils.toBuffer(code), function(err) {
+        if (err) {
+          return callback(err);
+        }
+        callback(null, account);
+      });
+    }
+  );
 };
 
-ForkedBlockchain.prototype.fetchAccountFromFallback = function(address, block_number, callback) {
+ForkedBlockchain.prototype.fetchCodeFromFallback = function(address, blockNumber, callback) {
   var self = this;
   address = to.hex(address);
 
-  async.parallel({
-    code: this.fetchCodeFromFallback.bind(this, address, block_number),
-    balance: this.fetchBalanceFromFallback.bind(this, address, block_number),
-    nonce: this.fetchNonceFromFallback.bind(this, address, block_number)
-  }, function(err, results) {
-    if (err) return callback(err);
-
-    var code = results.code;
-    var balance = results.balance;
-    var nonce = results.nonce;
-
-    var account = new Account({
-      nonce: nonce,
-      balance: balance
-    });
-
-    account.exists = code != "0x0" || balance != "0x0" || nonce != "0x0";
-
-    // This puts the code on the trie, keyed by the hash of the code.
-    // It does not actually link an account to code in the trie.
-    account.setCode(self.stateTrie, utils.toBuffer(code), function(err) {
-      if (err) return callback(err);
-      callback(null, account);
-    });
-  });
-};
-
-ForkedBlockchain.prototype.fetchCodeFromFallback = function(address, block_number, callback) {
-  var self = this;
-  address = to.hex(address);
-
-  // Allow an optional block_number
-  if (typeof block_number == "function") {
-    callback = block_number;
-    block_number = this.fork_block_number;
+  // Allow an optional blockNumber
+  if (typeof blockNumber === "function") {
+    callback = blockNumber;
+    blockNumber = this.forkBlockNumber;
   }
 
-  this.getSafeFallbackBlockNumber(block_number, function(err, safe_block_number) {
-    if (err) return callback(err);
+  this.getSafeFallbackBlockNumber(blockNumber, function(err, safeBlockNumber) {
+    if (err) {
+      return callback(err);
+    }
 
-    self.web3.eth.getCode(address, safe_block_number, function(err, code) {
-      if (err) return callback(err);
+    self.web3.eth.getCode(address, safeBlockNumber, function(err, code) {
+      if (err) {
+        return callback(err);
+      }
 
       code = "0x" + utils.toBuffer(code).toString("hex");
       callback(null, code);
     });
   });
-}
+};
 
-ForkedBlockchain.prototype.fetchBalanceFromFallback = function(address, block_number, callback) {
+ForkedBlockchain.prototype.fetchBalanceFromFallback = function(address, blockNumber, callback) {
   var self = this;
   address = to.hex(address);
 
-  // Allow an optional block_number
-  if (typeof block_number == "function") {
-    callback = block_number;
-    block_number = this.fork_block_number;
+  // Allow an optional blockNumber
+  if (typeof blockNumber === "function") {
+    callback = blockNumber;
+    blockNumber = this.forkBlockNumber;
   }
 
-  this.getSafeFallbackBlockNumber(block_number, function(err, safe_block_number) {
-    if (err) return callback(err);
+  this.getSafeFallbackBlockNumber(blockNumber, function(err, safeBlockNumber) {
+    if (err) {
+      return callback(err);
+    }
 
-    self.web3.eth.getBalance(address, safe_block_number, function(err, balance) {
-      if (err) return callback(err);
+    self.web3.eth.getBalance(address, safeBlockNumber, function(err, balance) {
+      if (err) {
+        return callback(err);
+      }
 
-      balance = "0x" + balance.toString(16); // BigNumber
+      balance = "0x" + new BN(balance).toString(16);
       callback(null, balance);
     });
   });
-}
+};
 
-ForkedBlockchain.prototype.fetchNonceFromFallback = function(address, block_number, callback) {
+ForkedBlockchain.prototype.fetchNonceFromFallback = function(address, blockNumber, callback) {
   var self = this;
   address = to.hex(address);
 
-  // Allow an optional block_number
-  if (typeof block_number == "function") {
-    callback = block_number;
-    block_number = this.fork_block_number;
+  // Allow an optional blockNumber
+  if (typeof blockNumber === "function") {
+    callback = blockNumber;
+    blockNumber = this.forkBlockNumber;
   }
 
-  this.getSafeFallbackBlockNumber(block_number, function(err, safe_block_number) {
-    if (err) return callback(err);
+  this.getSafeFallbackBlockNumber(blockNumber, function(err, safeBlockNumber) {
+    if (err) {
+      return callback(err);
+    }
 
-    self.web3.eth.getTransactionCount(address, safe_block_number, function(err, nonce) {
-      if (err) return callback(err);
+    self.web3.eth.getTransactionCount(address, safeBlockNumber, function(err, nonce) {
+      if (err) {
+        return callback(err);
+      }
 
-      nonce = "0x" + self.web3.toBigNumber(nonce).toString(16);
+      nonce = "0x" + self.web3.utils.toBN(nonce).toString(16);
       callback(null, nonce);
     });
   });
-}
+};
 
 ForkedBlockchain.prototype.getHeight = function(callback) {
-  var self = this;
   this.latestBlock(function(err, block) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
     callback(null, to.number(block.header.number));
   });
 };
@@ -506,21 +579,26 @@ ForkedBlockchain.prototype.getHeight = function(callback) {
 ForkedBlockchain.prototype.getRelativeBlockNumber = function(number, callback) {
   var self = this;
   this.getEffectiveBlockNumber(number, function(err, effective) {
-    if (err) return callback(err);
-    callback(null, effective - to.number(self.fork_block_number) - 1)
+    if (err) {
+      return callback(err);
+    }
+    callback(null, effective - to.number(self.forkBlockNumber) - 1);
   });
 };
 
-ForkedBlockchain.prototype.getSafeFallbackBlockNumber = function(block_number, callback) {
-  var fork_block_number = to.number(this.fork_block_number);
+ForkedBlockchain.prototype.getSafeFallbackBlockNumber = function(blockNumber, callback) {
+  var forkBlockNumber = to.number(this.forkBlockNumber);
 
-  if (block_number == null) {
-    return callback(null, fork_block_number);
-  };
+  if (blockNumber == null) {
+    return callback(null, forkBlockNumber);
+  }
 
-  this.getEffectiveBlockNumber(block_number, function(err, effective) {
-    if (effective > fork_block_number) {
-      effective = fork_block_number
+  this.getEffectiveBlockNumber(blockNumber, function(err, effective) {
+    if (err) {
+      return callback(err);
+    }
+    if (effective > forkBlockNumber) {
+      effective = forkBlockNumber;
     }
 
     callback(null, effective);
@@ -531,68 +609,52 @@ ForkedBlockchain.prototype.getBlockLogs = function(number, callback) {
   var self = this;
 
   this.getEffectiveBlockNumber(number, function(err, effective) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     self.getRelativeBlockNumber(effective, function(err, relative) {
-      if (err) return callback(err);
+      if (err) {
+        return callback(err);
+      }
 
       if (relative < 0) {
         self.getBlock(number, function(err, block) {
-          if (err) return callback(err);
+          if (err) {
+            return callback(err);
+          }
 
-          self.web3.currentProvider.sendAsync({
-            jsonrpc: "2.0",
-            method: "eth_getLogs",
-            params: [{
-              fromBlock: to.hex(number),
-              toBlock: to.hex(number)
-            }],
-            id: new Date().getTime()
-          }, function(err, res) {
-            if (err) return callback(err);
+          self.web3.currentProvider.send(
+            {
+              jsonrpc: "2.0",
+              method: "eth_getLogs",
+              params: [
+                {
+                  fromBlock: to.hex(number),
+                  toBlock: to.hex(number)
+                }
+              ],
+              id: new Date().getTime()
+            },
+            function(err, res) {
+              if (err) {
+                return callback(err);
+              }
 
-            var logs = res.result.map(function(log) {
-              // To make this result masquerade as the right information.
-              log.block = block;
-              return new Log(log);
-            });
+              var logs = res.result.map(function(log) {
+                // To make this result masquerade as the right information.
+                log.block = block;
+                return new Log(log);
+              });
 
-            callback(null, logs);
-          });
+              callback(null, logs);
+            }
+          );
         });
       } else {
         BlockchainDouble.prototype.getBlockLogs.call(self, relative, callback);
       }
     });
-  });
-};
-
-ForkedBlockchain.prototype._checkpointTrie = function() {
-  var self = this;
-
-  BlockchainDouble.prototype._checkpointTrie.call(this);
-
-  Object.keys(this.storageTrieCache).forEach(function(address) {
-    var trie = self.storageTrieCache[address];
-    trie.customCheckpoint();
-  });
-};
-
-ForkedBlockchain.prototype._revertTrie = function() {
-  var self = this;
-
-  BlockchainDouble.prototype._revertTrie.call(this);
-
-  Object.keys(this.storageTrieCache).forEach(function(address) {
-    var trie = self.storageTrieCache[address];
-
-    // We're trying to revert to a point before this trie was created.
-    // Let's just remove the trie.
-    if (trie.checkpoints.length == 0) {
-      delete self.storageTrieCache[address];
-    } else {
-      trie.customRevert();
-    }
   });
 };
 
