@@ -1,77 +1,71 @@
-var MerklePatriciaTree = require("merkle-patricia-tree");
-var Account = require("ethereumjs-account");
-var utils = require('ethereumjs-util')
+const MerklePatriciaTree = require("merkle-patricia-tree");
+const BaseTrie = require("merkle-patricia-tree/baseTrie");
+const checkpointInterface = require("merkle-patricia-tree/checkpoint-interface");
+var utils = require("ethereumjs-util");
 var inherits = require("util").inherits;
 var Web3 = require("web3");
 var to = require("./to.js");
-var async = require("async");
 
-inherits(ForkedStorageTrie, MerklePatriciaTree)
+inherits(ForkedStorageBaseTrie, BaseTrie);
 
-function ForkedStorageTrie(db, root, options) {
-  MerklePatriciaTree.call(this, db, root);
+function ForkedStorageBaseTrie(db, root, options) {
+  BaseTrie.call(this, db, root);
 
+  this.options = options;
   this.address = options.address;
-
-  this.fork = options.fork;
-  this.fork_block_number = options.fork_block_number;
-
+  this.forkBlockNumber = options.forkBlockNumber;
   this.blockchain = options.blockchain;
-
-  this.web3 = new Web3();
-  this.web3.setProvider(this.fork);
-
-  this.checkpoints = [];
+  this.fork = options.fork;
+  this.web3 = new Web3(this.fork);
 }
 
-ForkedStorageTrie.prototype.keyExists = function(key, callback) {
-  key = utils.toBuffer(key);
-
-  this.findPath(key, function (err, node, remainder, stack) {
-    var exists = false;
-    if (node && remainder.length === 0) {
-      exists = true;
-    }
-    callback(err, exists)
-  })
-};
-
 // Note: This overrides a standard method whereas the other methods do not.
-ForkedStorageTrie.prototype.get = function(key, block_number, callback) {
+ForkedStorageBaseTrie.prototype.get = function(key, blockNumber, callback) {
   var self = this;
 
-  // Allow an optional block_number
-  if (typeof block_number == "function") {
-    callback = block_number;
-    block_number = self.fork_block_number;
+  // Allow an optional blockNumber
+  if (typeof blockNumber === "function") {
+    callback = blockNumber;
+    blockNumber = this.forkBlockNumber;
   }
-
-  // For geth; https://github.com/ethereumjs/ethereumjs-util/issues/79
-  block_number = to.hexWithoutLeadingZeroes(block_number);
 
   key = utils.toBuffer(key);
 
-  // If the account doesn't exist in our state trie, get it off the wire.
+  // If the account/key doesn't exist in our state trie, get it off the wire.
   this.keyExists(key, function(err, exists) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     if (exists) {
+      // TODO: just because we have the key doesn't mean we're at the right
+      // block number/root to send it. We need to check the block number
+      // before using the data in our own trie.
       MerklePatriciaTree.prototype.get.call(self, key, function(err, r) {
         callback(err, r);
       });
     } else {
       // If this is the main trie, get the whole account.
       if (self.address == null) {
-        self.blockchain.fetchAccountFromFallback(key, block_number, function(err, account) {
-          if (err) return callback(err);
+        self.blockchain.fetchAccountFromFallback(key, blockNumber, function(err, account) {
+          if (err) {
+            return callback(err);
+          }
 
           callback(null, account.serialize());
         });
       } else {
-        self.web3.eth.getStorageAt(to.hex(self.address), to.hex(key), block_number, function(err, value) {
-          if (err) return callback(err);
+        if (to.number(blockNumber) > to.number(self.forkBlockNumber)) {
+          blockNumber = self.forkBlockNumber;
+        }
+        self.web3.eth.getStorageAt(to.rpcDataHexString(self.address), to.rpcDataHexString(key), blockNumber, function(
+          err,
+          value
+        ) {
+          if (err) {
+            return callback(err);
+          }
 
-          value = utils.toBuffer(value);
           value = utils.rlp.encode(value);
 
           callback(null, value);
@@ -81,16 +75,27 @@ ForkedStorageTrie.prototype.get = function(key, block_number, callback) {
   });
 };
 
-// I don't want checkpoints to be removed by commits.
-// Note: For some reason, naming this function checkpoint()
-// -- overriding the default function -- prevents it from
-// being called.
-ForkedStorageTrie.prototype.customCheckpoint = function() {
-  this.checkpoints.push(this.root);
+ForkedStorageBaseTrie.prototype.keyExists = function(key, callback) {
+  key = utils.toBuffer(key);
+
+  this.findPath(key, function(err, node, remainder, stack) {
+    const exists = node && remainder.length === 0;
+    callback(err, exists);
+  });
 };
 
-ForkedStorageTrie.prototype.customRevert = function() {
-  this.root = this.checkpoints.pop();
+ForkedStorageBaseTrie.prototype.copy = function() {
+  return new ForkedStorageBaseTrie(this.db, this.root, this.options);
 };
+
+inherits(ForkedStorageTrie, ForkedStorageBaseTrie);
+
+function ForkedStorageTrie(db, root, options) {
+  ForkedStorageBaseTrie.call(this, db, root, options);
+  checkpointInterface(this);
+}
+
+ForkedStorageTrie.prove = MerklePatriciaTree.prove;
+ForkedStorageTrie.verifyProof = MerklePatriciaTree.verifyProof;
 
 module.exports = ForkedStorageTrie;
