@@ -19,14 +19,17 @@ function Log(data) {
 }
 
 Log.prototype.toJSON = function() {
+  // RPC quantity values like this.transactionIndex can be set to "0x00",
+  // use the explicit rpcQuantityHexString to properly format the JSON, removing leading zeroes.
+  // See RPC log format spec: https://github.com/ethereum/wiki/wiki/JSON-RPC
   return {
-    logIndex: this.logIndex,
-    transactionIndex: to.hex(this.transactionIndex),
-    transactionHash: to.hex(this.transactionHash),
-    blockHash: to.hex(this.block.hash()),
-    blockNumber: to.hex(this.block.header.number),
-    address: to.hex(this.address),
-    data: to.hex(this.data),
+    logIndex: to.rpcQuantityHexString(this.logIndex),
+    transactionIndex: to.rpcQuantityHexString(this.transactionIndex),
+    transactionHash: to.rpcDataHexString(this.transactionHash),
+    blockHash: to.rpcDataHexString(this.block.hash()),
+    blockNumber: to.rpcQuantityHexString(this.block.header.number),
+    address: to.rpcDataHexString(this.address),
+    data: to.rpcDataHexString(this.data),
     topics: this.topics,
     type: "mined"
   };
