@@ -1,18 +1,22 @@
 var to = require("./to");
 
-function Receipt(tx, block, logs, gasUsed, contractAddress, status) {
+function Receipt(tx, block, logs, gasUsed, cumulativeGasUsed, contractAddress, status, logsBloom) {
   this.tx = tx;
   this.block = block;
   this.logs = logs;
   this.gasUsed = gasUsed;
+  this.cumulativeGasUsed = cumulativeGasUsed;
   this.contractAddress = contractAddress;
   this.status = status;
+  this.logsBloom = logsBloom;
 
   this.transactionIndex = 0;
 
+  this.txHash = tx.hash();
+
   for (var i = 0; i < block.transactions.length; i++) {
     var current = block.transactions[i];
-    if (current.hash().equals(tx.hash())) {
+    if (current.hash().equals(this.txHash)) {
       this.transactionIndex = i;
       break;
     }
@@ -20,19 +24,26 @@ function Receipt(tx, block, logs, gasUsed, contractAddress, status) {
 }
 
 Receipt.prototype.toJSON = function() {
-  if (this.data != null) return data;
-
+  // Enforce Hex formatting as defined in the RPC spec.
   return {
-    transactionHash: to.hex(this.tx.hash()),
-    transactionIndex: to.hex(this.transactionIndex),
-    blockHash: to.hex(this.block.hash()),
-    blockNumber: to.hex(this.block.header.number),
-    gasUsed: to.hex(this.gasUsed),
-    cumulativeGasUsed: to.hex(this.block.header.gasUsed),
-    contractAddress: this.contractAddress != null ? to.hex(this.contractAddress) : null,
-    logs: this.logs.map(function(log) {return log.toJSON()}),
-    status: to.number(this.status)
-  }
+    transactionHash: to.rpcDataHexString(this.txHash),
+    transactionIndex: to.rpcQuantityHexString(this.transactionIndex),
+    blockHash: to.rpcDataHexString(this.block.hash()),
+    blockNumber: to.rpcQuantityHexString(this.block.header.number),
+    from: to.rpcDataHexString(this.tx.from),
+    to: to.nullableRpcDataHexString(this.tx.to),
+    gasUsed: to.rpcQuantityHexString(this.gasUsed),
+    cumulativeGasUsed: to.rpcQuantityHexString(this.cumulativeGasUsed),
+    contractAddress: this.contractAddress != null ? to.rpcDataHexString(this.contractAddress) : null,
+    logs: this.logs.map(function(log) {
+      return log.toJSON();
+    }),
+    status: to.rpcQuantityHexString(this.status),
+    logsBloom: to.rpcDataHexString(this.logsBloom),
+    v: to.rpcQuantityHexString(this.tx.v),
+    r: to.rpcQuantityHexString(this.tx.r),
+    s: to.rpcQuantityHexString(this.tx.s)
+  };
 };
 
 module.exports = Receipt;
