@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 var TestRPC = require("../");
 var Web3 = require("web3");
-var async = require("async")
+var async = require("async");
 
 var server = TestRPC.server();
 var port = 12345;
@@ -11,7 +11,9 @@ function runTest(times, fn, callback) {
   var start = new Date();
 
   async.timesSeries(times, fn, function(err) {
-    if (err) return callback(err);
+    if (err) {
+      return callback(err);
+    }
 
     var end = new Date();
     var actualTime = end.getTime() - start.getTime();
@@ -20,37 +22,44 @@ function runTest(times, fn, callback) {
   });
 }
 
-function runAverage(title, number_of_runs, fn_times, fn, callback) {
-  var results = new Array(number_of_runs);
+function runAverage(title, numberOfRuns, fnTimes, fn, callback) {
+  var results = new Array(numberOfRuns);
 
-  async.timesSeries(number_of_runs, function(n, next) {
-    process.stdout.write(title + " " + (n + 1) + "...");
+  async.timesSeries(
+    numberOfRuns,
+    function(n, next) {
+      process.stdout.write(title + " " + (n + 1) + "...");
 
-    runTest(fn_times, fn, function(err, totalTime) {
-      if (err) return next(err);
-      results[n] = totalTime;
+      runTest(fnTimes, fn, function(err, totalTime) {
+        if (err) {
+          return next(err);
+        }
+        results[n] = totalTime;
 
-      console.log((totalTime / 1000) + " seconds");
-      next();
-    });
-  }, function(err) {
-    if (err) return callback(err);
+        console.log(totalTime / 1000 + " seconds");
+        next();
+      });
+    },
+    function(err) {
+      if (err) {
+        return callback(err);
+      }
 
-    var sum = results.reduce(function(a, b) {
-      return a + b;
-    }, 0);
+      var sum = results.reduce(function(a, b) {
+        return a + b;
+      }, 0);
 
-    var average = sum / number_of_runs;
+      var average = sum / numberOfRuns;
 
-    console.log("Average " + (average / 1000) + " seconds");
+      console.log("Average " + average / 1000 + " seconds");
 
-    callback(null, average);
-  });
-};
+      callback(null, average);
+    }
+  );
+}
 
-function bailIfError(err) {
-  if (err) {
-    console.log(err);
+function bailIfError(_) {
+  if (_) {
     process.exit(1);
   }
 }
@@ -61,16 +70,25 @@ server.listen(port, function(err) {
   web3.eth.getAccounts(function(err, accounts) {
     bailIfError(err);
 
-    runAverage("Running transactions test", 4, 1000, function(n, cb) {
-      web3.eth.sendTransaction({
-        from: accounts[0],
-        to: accounts[1],
-        value: 500, // wei
-        gas: 90000
-      }, cb);
-    }, function(err) {
-      bailIfError(err);
-      server.close();
-    });
+    runAverage(
+      "Running transactions test",
+      4,
+      1000,
+      function(n, cb) {
+        web3.eth.sendTransaction(
+          {
+            from: accounts[0],
+            to: accounts[1],
+            value: 500, // wei
+            gas: 90000
+          },
+          cb
+        );
+      },
+      function(err) {
+        bailIfError(err);
+        server.close();
+      }
+    );
   });
 });
