@@ -1,6 +1,9 @@
-var Web3 = require('web3');
-var TestRPC = require("../index.js");
-var assert = require('assert');
+var Web3 = require("web3");
+var Ganache = require(process.env.TEST_BUILD
+  ? "../build/ganache.core." + process.env.TEST_BUILD + ".js"
+  : "../index.js");
+var assert = require("assert-match");
+var regex = require("assert-match/matchers").regex;
 
 var tests = function(web3) {
   var accounts;
@@ -14,7 +17,9 @@ var tests = function(web3) {
   describe("bad input", function() {
     before(function(done) {
       web3.eth.getAccounts(function(err, accs) {
-        if (err) return done(err);
+        if (err) {
+          return done(err);
+        }
         accounts = accs;
         done();
       });
@@ -23,58 +28,72 @@ var tests = function(web3) {
     it("recovers after to address that isn't a string", function(done) {
       var provider = web3.currentProvider;
 
-      provider.sendAsync({
-        "jsonrpc": "2.0",
-        "method": "eth_sendTransaction",
-        "params": [
-          {
-            "value": "0x0",
-            "gas": "0xf4240",
-            "from": accounts[0],
-            // Buffers have been sent in the past
-            "to": {
-              "type": "Buffer",
-              "data": [
-                // ...
-              ]
-            },
-            "data": "0xe1fa8e84666f6f0000000000000000000000000000000000000000000000000000000000"
-          }
-        ],
-        "id": 2
-      }, function() {
-        // Ignore any errors, but make sure we can make the second request
-        secondRequest(done);
-      });
+      provider.send(
+        {
+          jsonrpc: "2.0",
+          method: "eth_sendTransaction",
+          params: [
+            {
+              value: "0x0",
+              gas: "0xf4240",
+              from: accounts[0],
+              // Buffers have been sent in the past
+              to: {
+                type: "Buffer",
+                data: [
+                  // ...
+                ]
+              },
+              data: "0xe1fa8e84666f6f0000000000000000000000000000000000000000000000000000000000"
+            }
+          ],
+          id: 2
+        },
+        function() {
+          // Ignore any errors, but make sure we can make the second request
+          secondRequest(done);
+        }
+      );
     });
 
     it("recovers after bad nonce (too high)", function(done) {
       var provider = web3.currentProvider;
 
       var request = {
-        "jsonrpc": "2.0",
-        "method": "eth_sendTransaction",
-        "params": [
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
           {
-            "value": "0x10000000",
-            "gas": "0xf4240",
-            "from": accounts[0],
-            "to": accounts[1],
-            "nonce": "0xffffffff",  // too big nonce
+            value: "0x10000000",
+            gas: "0xf4240",
+            from: accounts[0],
+            to: accounts[1],
+            nonce: "0xffffffff" // too big nonce
           }
         ],
-        "id": 2
-      }
+        id: 2
+      };
 
-      provider.sendAsync(request, function(err, result) {
+      provider.send(request, function(err, result) {
+        if (err) {
+          assert(
+            err.message.indexOf(
+              "the tx doesn't have the correct nonce. account has nonce of: 0 tx has nonce of: 4294967295"
+            ) >= 0
+          );
+        }
         // We're supposed to get an error the first time. Let's assert we get the right one.
-        // Note that if using the TestRPC as a provider, err will be non-null when there's
+        // Note that if using the Ganache as a provider, err will be non-null when there's
         // an error. However, when using it as a server it won't be. In both cases, however,
         // result.error should be set with the same error message. We'll check for that.
-        assert(result.error.message.indexOf("the tx doesn't have the correct nonce. account has nonce of: 0 tx has nonce of: 4294967295") >= 0);
+        assert(
+          result.error.message.indexOf(
+            "the tx doesn't have the correct nonce. account has nonce of: 0 tx has nonce of: 4294967295"
+          ) >= 0
+        );
 
-        delete request.params[0].nonce
-        provider.sendAsync(request, done)
+        delete request.params[0].nonce;
+        provider.send(request, done);
       });
     });
 
@@ -82,70 +101,94 @@ var tests = function(web3) {
       var provider = web3.currentProvider;
 
       var request = {
-        "jsonrpc": "2.0",
-        "method": "eth_sendTransaction",
-        "params": [
+        jsonrpc: "2.0",
+        method: "eth_sendTransaction",
+        params: [
           {
-            "value": "0x10000000",
-            "gas": "0xf4240",
-            "from": accounts[0],
-            "to": accounts[1],
-            "nonce": "0x00",  // too low nonce
+            value: "0x10000000",
+            gas: "0xf4240",
+            from: accounts[0],
+            to: accounts[1],
+            nonce: "0x0" // too low nonce
           }
         ],
-        "id": 2
-      }
+        id: 2
+      };
 
-      provider.sendAsync(request, function(err, result) {
+      provider.send(request, function(err, result) {
+        if (err) {
+          assert(
+            /the tx doesn't have the correct nonce. account has nonce of: 1 tx has nonce of: 0/.test(err.message),
+            `Expected incorrect nonce error, got '${err.message}', instead.`
+          );
+        }
         // We're supposed to get an error the first time. Let's assert we get the right one.
-        // Note that if using the TestRPC as a provider, err will be non-null when there's
+        // Note that if using the Ganache as a provider, err will be non-null when there's
         // an error. However, when using it as a server it won't be. In both cases, however,
         // result.error should be set with the same error message. We'll check for that.
-        assert(result.error.message.indexOf("the tx doesn't have the correct nonce. account has nonce of: 1 tx has nonce of: 0") >= 0);
+        assert(
+          /the tx doesn't have the correct nonce. account has nonce of: 1 tx has nonce of: 0/.test(
+            result.error.message
+          ),
+          `Expected incorrect nonce error, got '${result.error.message}', instead.`
+        );
 
-        delete request.params[0].nonce
-        provider.sendAsync(request, done)
+        delete request.params[0].nonce;
+        provider.send(request, done);
       });
     });
 
     it("recovers after bad balance", function(done) {
-      web3.eth.getBalance(accounts[0], function(err, balance) {
-        if (err) return done(err);
-
+      web3.eth.getBalance(accounts[0], function(_, balance) {
         var provider = web3.currentProvider;
 
         var request = {
-          "jsonrpc": "2.0",
-          "method": "eth_sendTransaction",
-          "params": [
+          jsonrpc: "2.0",
+          method: "eth_sendTransaction",
+          params: [
             {
-              "value": "0x1000000000000000000000000000",
-              "gas": "0xf4240",
-              "from": accounts[0],
-              "to": accounts[1]
+              value: "0x1000000000000000000000000000",
+              gas: "0xf4240",
+              from: accounts[0],
+              to: accounts[1]
             }
           ],
-          "id": 2
-        }
+          id: 2
+        };
 
-        provider.sendAsync(request, function(err, result) {
+        provider.send(request, function(err, result) {
+          if (err) {
+            assert.deepEqual(
+              err.message,
+              regex(
+                /sender doesn't have enough funds to send tx. The upfront cost is: \d+ and the sender's account only has: \d+/
+              ),
+              `Unexpected error message. Got ${err.message}.`
+            );
+          }
           // We're supposed to get an error the first time. Let's assert we get the right one.
-          // Note that if using the TestRPC as a provider, err will be non-null when there's
+          // Note that if using the Ganache as a provider, err will be non-null when there's
           // an error. However, when using it as a server it won't be. In both cases, however,
           // result.error should be set with the same error message. We'll check for that.
-          assert(result.error.message.indexOf("sender doesn't have enough funds to send tx. The upfront cost is: 324518553658426726783156021576256 and the sender's account only has: 99999999999463087088") >= 0);
+          assert.deepEqual(
+            result.error.message,
+            regex(
+              /sender doesn't have enough funds to send tx. The upfront cost is: \d+ and the sender's account only has: \d+/
+            ),
+            `Unexpected error message. Got ${result.error.message}.`
+          );
 
           request.params[0].value = "0x5";
-          provider.sendAsync(request, done)
+          provider.send(request, done);
         });
-      })
+      });
     });
-  })
+  });
 };
 
 describe("Provider:", function() {
   var web3 = new Web3();
-  web3.setProvider(TestRPC.provider());
+  web3.setProvider(Ganache.provider({}));
   tests(web3);
 });
 
@@ -154,8 +197,8 @@ describe("Server:", function(done) {
   var port = 12345;
   var server;
 
-  before("Initialize TestRPC server", function(done) {
-    server = TestRPC.server();
+  before("Initialize Ganache server", function(done) {
+    server = Ganache.server({});
     server.listen(port, function() {
       web3.setProvider(new Web3.providers.HttpProvider("http://localhost:" + port));
       done();
