@@ -1,307 +1,293 @@
-var Web3 = require('web3');
-var TestRPC = require("../index.js");
-var assert = require('assert');
-var solc = require("solc");
-var async = require("async");
-
-var source = "                      \
-pragma solidity ^0.4.2;             \
-contract EventTest {                \
-  event ExampleEvent(uint indexed first, uint indexed second);   \
-                                    \
-  function triggerEvent(uint _first, uint _second) public { \
-    ExampleEvent(_first, _second);      \
-  }                                 \
-}"
-
-// Thanks solc. At least this works!
-// This removes solc's overzealous uncaughtException event handler.
-process.removeAllListeners("uncaughtException");
+var Web3 = require("web3");
+var Web3WsProvider = require("web3-providers-ws");
+var Ganache = require(process.env.TEST_BUILD
+  ? "../build/ganache.core." + process.env.TEST_BUILD + ".js"
+  : "../index.js");
+var assert = require("assert");
+const compile = require("./helpers/contract/singleFileCompile");
 
 var tests = function(web3, EventTest) {
   var accounts;
-  var EventTest;
   var instance;
 
   describe("events", function() {
     before(function(done) {
       web3.eth.getAccounts(function(err, accs) {
-        if (err) return done(err);
+        if (err) {
+          return done(err);
+        }
         accounts = accs;
         done();
       });
     });
 
-    before(function() {
-      var result = solc.compile(source, 1);
+    before(function(done) {
+      this.timeout(10000);
+      var { result } = compile("./test/contracts/events/", "EventTest");
 
       if (result.errors != null) {
-        throw new Error(result.errors[0]);
+        done(result.errors[0]);
+        return;
       }
 
-      var abi = JSON.parse(result.contracts[":EventTest"].interface);
-      EventTest = web3.eth.contract(abi);
-      EventTest._data = "0x" + result.contracts[":EventTest"].bytecode;
+      var abi = result.contracts["EventTest.sol"].EventTest.abi;
+      EventTest = new web3.eth.Contract(abi);
+      EventTest._data = "0x" + result.contracts["EventTest.sol"].EventTest.evm.bytecode.object;
+      done();
     });
 
-    before(function(done) {
-      EventTest.new({from: accounts[0], data: EventTest._data, gas: 3141592}, function(err, contract) {
-        if (err) return done(err);
+    before(function() {
+      return EventTest.deploy({ data: EventTest._data })
+        .send({ from: accounts[0], gas: 3141592 })
+        .then((contract) => {
+          instance = contract;
 
-        if (!contract.address) {
-          return;
-        }
-        instance = contract;
+          // TODO: ugly workaround - not sure why this is necessary.
+          if (!instance._requestManager.provider) {
+            instance._requestManager.setProvider(web3.eth._provider);
+          }
+        });
+    });
+
+    it("should handle events properly via the data event handler", function(done) {
+      var expectedValue = "1";
+
+      var event = instance.events.ExampleEvent({ filter: { first: expectedValue } });
+
+      var listener = function(result) {
+        assert.strictEqual(result.returnValues.first, expectedValue);
         done();
-      });
-    });
-
-    it("handles events properly, using `event.watch()`", function(done) {
-      var expected_value = 5;
-
-      var event = instance.ExampleEvent({first: expected_value});
-
-      var cleanup = function(err) {
-        if (err) return done(err);
-        event.stopWatching(done);
       };
 
-      event.watch(function(err, result) {
-        if (err) return cleanup(err);
+      event.once("data", listener);
+      event.once("error", (err) => done(err));
 
-        if (result.args.first == expected_value) {
-          return cleanup();
-        }
-
-        return cleanup(new Error("Received event that didn't have the correct value!"));
-      });
-
-      instance.triggerEvent(5, 6, {from: accounts[0], gas: 3141592}, function(err, result) {
-        if (err) return cleanup(err);
-      });
-    });
-
-    it("handles events properly, using `event.get()`", function(done) {
-      this.timeout(10000)
-      var expected_value = 6;
-      var interval;
-
-      var event = instance.ExampleEvent({first: expected_value});
-
-      function cleanup(err) {
-        if (err) return done(err);
-
-        event.stopWatching(function(err) {
-          if (err) return done(err);
-          clearInterval(interval);
-          done(err);
-        });
-      }
-
-      instance.triggerEvent(6, 7, {from: accounts[0]}, function(err, result) {
-        if (err) return cleanup(err);
-
-        interval = setInterval(function() {
-          event.get(function(err, logs) {
-            if (err) return cleanup(err);
-
-            if (logs.length == 0) return;
-
-            if (logs[0].args.first == expected_value) {
-              return cleanup();
-            }
-
-            return cleanup(new Error("Received event that didn't have the correct value!"));
-          });
-        }, 500);
-      });
+      instance.methods.triggerEvent(1, 6).send({ from: accounts[0], gas: 3141592 });
     });
 
     // NOTE! This test relies on the events triggered in the tests above.
-    it("grabs events in the past, using `event.get()`", function(done) {
-      var expected_value = 5;
-      var event = instance.ExampleEvent({first: expected_value}, {fromBlock: 0});
+    it("grabs events in the past", function(done) {
+      var expectedValue = "2";
 
-      event.get(function(err, logs) {
-        if (err) return done(err);
+      var event = instance.events.ExampleEvent({ filter: { first: expectedValue }, fromBlock: 0 });
 
-        event.stopWatching(function(err) {
-          if (err) return done(err);
-          assert(logs.length == 1);
-          done();
-        });
-      });
+      var listener = function(result) {
+        assert.strictEqual(result.returnValues.first, expectedValue);
+        done();
+      };
+
+      event.once("data", listener);
+
+      instance.methods.triggerEvent(2, 6).send({ from: accounts[0], gas: 3141592 });
     });
 
     // NOTE! This test relies on the events triggered in the tests above.
     it("accepts an array of topics as a filter", function(done) {
-      var expected_value_a = 5;
-      var expected_value_b = 6;
-      var event = instance.ExampleEvent({first: [expected_value_a, expected_value_b]}, {fromBlock: 0});
+      var expectedValueA = 3;
+      var expectedValueB = 4;
 
-      event.get(function(err, logs) {
-        if (err) return done(err);
+      var event = instance.events.ExampleEvent({ filter: { first: [expectedValueA, expectedValueB] }, fromBlock: 0 });
 
-        event.stopWatching(function(err) {
-          if (err) return done(err);
+      var waitingFor = {};
+      waitingFor[expectedValueA] = true;
+      waitingFor[expectedValueB] = true;
 
-          assert(logs.length == 2);
+      var listener = function(result) {
+        assert(waitingFor.hasOwnProperty(result.returnValues.first));
+        delete waitingFor[result.returnValues.first];
+
+        if (Object.keys(waitingFor).length === 0) {
+          event.removeAllListeners();
           done();
-        });
+        }
+      };
+
+      event.on("data", listener);
+
+      event.once("error", (err) => {
+        event.removeAllListeners();
+        done(err);
       });
+
+      instance.methods
+        .triggerEvent(expectedValueA, 6)
+        .send({ from: accounts[0], gas: 3141592 })
+        .then((result) => {
+          return instance.methods.triggerEvent(expectedValueB, 7).send({ from: accounts[0], gas: 3141592 });
+        });
     });
 
     it("only returns logs for the expected address", function(done) {
-      var expected_value = 5;
+      var expectedValue = "1";
 
-      EventTest.new({from: accounts[0], data: EventTest._data, gas: 3141592}, function(err, newInstance) {
-        if (err) return done(err);
+      EventTest.deploy({ data: EventTest._data })
+        .send({ from: accounts[0], gas: 3141592 })
+        .then((newInstance) => {
+          // TODO: ugly workaround - not sure why this is necessary.
+          if (!newInstance._requestManager.provider) {
+            newInstance._requestManager.setProvider(web3.eth._provider);
+          }
 
-        if (!newInstance.address) {
-          return;
-        }
+          var event = newInstance.events.ExampleEvent({ filter: { first: expectedValue }, fromBlock: 0 });
 
-        newInstance.triggerEvent(expected_value, 20, {from: accounts[0], gas: 3141592}, function(err, result) {
-          if (err) return done(err);
-
-          var event = newInstance.ExampleEvent({first: expected_value}, {fromBlock: 0});
-
-          // Only one event should be triggered for this new instance.
-          event.get(function(err, logs) {
-            if (err) return done(err);
-
-            event.stopWatching(function(err) {
-              if (err) return done(err);
-              assert(logs.length == 1);
-              done();
-            });
+          event.on("data", function(result) {
+            assert(result.returnValues.first === expectedValue);
+            // event.removeAllListeners()
+            done();
           });
+
+          instance.methods
+            .triggerEvent(5, 6)
+            .send({ from: accounts[0], gas: 3141592 })
+            .then(() => {
+              newInstance.methods.triggerEvent(expectedValue, 6).send({ from: accounts[0], gas: 3141592 });
+            });
         });
-      });
     });
 
-    it("always returns a change for every new block filter when instamining", function(done) {
+    // NOTE! This test relies on the events triggered in the tests above.
+    it("should return logs with correctly formatted logIndex and transactionIndex", function(done) {
       var provider = web3.currentProvider;
 
-      // In this test, we'll create a block filter and request filter changes twice.
-      // The responses from the first and second filter changes request must be different,
-      // and the first must return the block hash of the previous block to ensure it gets
-      // some response even though no transaction was made.
-
-      var filter_id;
-      var first_changes;
-      var second_changes;
-
-      async.series([
-        function(c) {
-          provider.sendAsync({
-            jsonrpc: "2.0",
-            method: "eth_newBlockFilter",
-            params: [],
-            id: new Date().getTime()
-          }, function(err, result) {
-            if (err) return c(err);
-            filter_id = result.result;
-            c();
-          });
+      provider.send(
+        {
+          jsonrpc: "2.0",
+          method: "eth_getLogs",
+          params: [
+            {
+              fromBlock: "0x0",
+              toBlock: "latest",
+              topics: [
+                "0xc54307031d9aa93e0568c363be84a9400dce343fef6a2851d55662a6af1a29da",
+                "0x0000000000000000000000000000000000000000000000000000000000000001",
+                "0x0000000000000000000000000000000000000000000000000000000000000006"
+              ]
+            }
+          ],
+          id: new Date().getTime()
         },
-        function(c) {
-          provider.sendAsync({
-            jsonrpc: "2.0",
-            method: "eth_getFilterChanges",
-            params: [filter_id],
-            id: new Date().getTime()
-          }, function(err, result) {
-            if (err) return c(err);
-            first_changes = result.result;
-            c();
-          });
-        },
-        function(c) {
-          provider.sendAsync({
-            jsonrpc: "2.0",
-            method: "eth_getFilterChanges",
-            params: [filter_id],
-            id: new Date().getTime()
-          }, function(err, result) {
-            if (err) return c(err);
-            second_changes = result.result;
-            c();
-          });
+        function(err, result) {
+          if (err) {
+            return done(err);
+          }
+          const logIndex = result.result[0].logIndex;
+          const transactionIndex = result.result[0].transactionIndex;
+          assert.strictEqual(logIndex, "0x0");
+          assert.strictEqual(transactionIndex, "0x0");
+          done();
         }
-      ], function(err) {
-        if (err) return done(err);
+      );
+    });
 
-        assert.equal(first_changes.length, 1);
-        assert.equal(first_changes[0].length, 66); // Ensure we have a hash
-        assert.equal(second_changes.length, 0); // no transactions were actually made
-        assert.notEqual(first_changes[0], second_changes[0]);
+    it("always returns a change for every new block subscription when instamining", function(done) {
+      var provider = web3.currentProvider;
 
-        done();
-      })
+      provider.send(
+        {
+          jsonrpc: "2.0",
+          method: "eth_subscribe",
+          params: ["newHeads"],
+          id: new Date().getTime()
+        },
+        function(err, result) {
+          if (err) {
+            return done(err);
+          }
+
+          let listener = function(err, result) {
+            if (result === undefined) {
+              // If there's only one argument, it's the result, not an error
+              result = err;
+            } else if (err) {
+              return done(err);
+            }
+            let firstChanges = result.params.result.hash;
+            assert.strictEqual(firstChanges.length, 66); // Ensure we have a hash
+            provider.removeAllListeners("data");
+            done();
+          };
+
+          // can't use `once` here because Web3WsProvider only has `on` :-(
+          provider.on("data", listener);
+
+          web3.currentProvider.send(
+            {
+              jsonrpc: "2.0",
+              method: "evm_mine",
+              id: new Date().getTime()
+            },
+            function(err) {
+              if (err) {
+                done(err);
+              }
+            }
+          );
+        }
+      );
     });
 
     // NOTE! This test relies on the events triggered in the tests above.
     it("ensures topics are respected in past events, using `event.get()` (exclusive)", function(done) {
-      var unexpected_value = 1337;
-      var event = instance.ExampleEvent({first: unexpected_value}, {fromBlock: 0});
+      var unexpectedValue = 1337;
+      var event = instance.events.ExampleEvent({ filter: { first: unexpectedValue }, fromBlock: 0 });
 
       // There should be no logs because we provided a different number.
-      event.get(function(err, logs) {
-        if (err) return done(err);
-        assert(logs.length == 0);
-        done();
-      });
+      var listener = function(result) {
+        assert.fail("Event should not have fired");
+      };
+
+      event.once("data", listener);
+
+      instance.methods
+        .triggerEvent(6, 6)
+        .send({ from: accounts[0], gas: 3141592 })
+        .then(() => {
+          // have to finish somehow...
+          setTimeout(() => {
+            event.removeAllListeners();
+            done();
+          }, 250);
+        });
     });
 
-    it("ensures topics are respected in past events, using `event.get()` (inclusive/exclusive)", function(done) {
-      var expected_value = 6;
-      var event = instance.ExampleEvent({second: expected_value}, {fromBlock: 0});
+    // TODO: web3 1.0 drops fromBlock on a subscription request - stop skipping this when that is fixed
+    it.skip("will not fire if logs are requested when fromBlock doesn't exist", function(done) {
+      var event = instance.events.ExampleEvent({ fromBlock: 100000 });
 
-      // There should be no logs because we provided a different number.
-      event.get(function(err, logs) {
-        if (err) return done(err);
-        assert(logs.length == 1);
-        done();
-      });
+      // fromBlock doesn't exist, hence no logs
+      var listener = function(result) {
+        assert.fail("Event should not have fired");
+      };
+
+      event.on("data", listener);
+
+      instance.methods
+        .triggerEvent(8, 6)
+        .send({ from: accounts[0], gas: 3141592 })
+        .then(() => {
+          // have to finish somehow...
+          setTimeout(() => {
+            event.removeAllListeners();
+            done();
+          }, 250);
+        });
     });
-
-    it("will return an empty array if logs are requested when fromBlock doesn't exist", function(done) {
-      var event = instance.ExampleEvent({}, {fromBlock: 100000});
-
-      // fromBlock doesn't exist, hence no logs.
-      event.get(function(err, logs) {
-        if (err) return done(err);
-        assert(logs.length == 0);
-        done();
-      });
-    });
-
-    it("will return an empty array if logs are requested when toBlock doesn't exist", function(done) {
-      var expected_value = 6;
-      var event = instance.ExampleEvent({}, {toBlock: 100000});
-
-      // fromBlock doesn't exist, hence no logs.
-      event.get(function(err, logs) {
-        if (err) return done(err);
-        assert(logs.length == 0);
-        done();
-      });
-    });
-  })
+  });
 };
 
 var logger = {
   log: function(message) {
-    //console.log(message);
+    // console.log(message);
   }
 };
 
 describe("Provider:", function() {
   var web3 = new Web3();
-  web3.setProvider(TestRPC.provider({
-    logger: logger
-  }));
+  web3.setProvider(
+    Ganache.provider({
+      logger: logger
+    })
+  );
   tests(web3);
 });
 
@@ -310,19 +296,21 @@ describe("Server:", function(done) {
   var port = 12345;
   var server;
 
-  before("Initialize TestRPC server", function(done) {
-    server = TestRPC.server({
-      logger: logger
+  before("Initialize Ganache server", function(done) {
+    server = Ganache.server({
+      logger: logger,
+      ws: true
     });
     server.listen(port, function() {
-      web3.setProvider(new Web3.providers.HttpProvider("http://localhost:" + port));
+      web3.setProvider(new Web3WsProvider("ws://localhost:" + port));
       done();
     });
   });
 
+  tests(web3);
+
   after("Shutdown server", function(done) {
+    web3._provider.connection.close();
     server.close(done);
   });
-
-  tests(web3);
 });
