@@ -1,83 +1,37 @@
-var Web3 = require('web3');
-var Transaction = require('ethereumjs-tx');
-var utils = require('ethereumjs-util');
-var assert = require('assert');
-var TestRPC = require("../index.js");
-var solc = require("solc");
-var fs = require("fs");
-var to = require("../lib/utils/to");
-var clone = require("clone");
-
-// Thanks solc. At least this works!
-// This removes solc's overzealous uncaughtException event handler.
-process.removeAllListeners("uncaughtException");
-
-var logger = {
-  log: function(message) {
-    //console.log(message);
-  }
-};
-
-var web3 = new Web3();
-web3.setProvider(TestRPC.provider({
-  /*blocktime: 100,*/
-  logger: logger,
-  seed: "1337"
-}));
+const assert = require("assert");
+const bootstrap = require("./helpers/contract/bootstrap");
 
 describe("revert opcode", function() {
-  var testContext = {};
-
-  before(function (done) {
+  let context;
+  before("Setting up web3 and contract", async function() {
     this.timeout(10000);
-    testContext.source = fs.readFileSync("./test/Revert.sol", {encoding: "utf8"});
-    testContext.solcResult = solc.compile(testContext.source, false);
-
-    testContext.revertContract = {
-      solidity: testContext.source,
-      abi: testContext.solcResult.contracts[":Revert"].interface,
-      binary: "0x" + testContext.solcResult.contracts[":Revert"].bytecode,
-      runtimeBinary: '0x' + testContext.solcResult.contracts[":Revert"].runtimeBytecode
+    const logger = {
+      log: function(message) {}
     };
 
-    web3.eth.getAccounts(function(err, accs) {
-      if (err) return done(err);
+    const contractRef = {
+      contractFiles: ["Revert"],
+      contractSubdirectory: "revert"
+    };
 
-      testContext.accounts = accs;
+    const ganacheProviderOptions = {
+      logger,
+      seed: "1337"
+    };
 
-      return done();
-    });
+    context = await bootstrap(contractRef, ganacheProviderOptions);
   });
 
-  it("should return a transaction receipt with status 0 on REVERT", function(done) {
-    var revertCode = testContext.revertContract.binary;
-    var revertAbi = JSON.parse(testContext.revertContract.abi);
-    var callCount = 0;
+  it("should return a transaction receipt with status 0 on REVERT", async function() {
+    const { accounts, instance, web3 } = context;
 
-    var RevertContract = web3.eth.contract(revertAbi);
-    RevertContract._code = revertCode;
-    RevertContract.new({ data: revertCode, from: testContext.accounts[0], gas: 3141592 }, function (err, instance) {
-      callCount++;
-      if (err) {
-        return done(err);
-      }
+    const promise = instance.methods.alwaysReverts(5).send({ from: accounts[0] });
+    const hash = await promise.catch((err) => err.hashes[0]);
+    const receipt = await web3.eth.getTransactionReceipt(hash);
 
-      if (instance.address) {
-        instance.alwaysReverts(5, { from: testContext.accounts[0] }, function(err, result) {
-          assert(err, "Expected error result not returned.");
-          var txHash = err.hashes[0];
+    await assert.rejects(promise, (err) => err.results[hash].error === "revert", "Expected error result not returned.");
 
-          web3.eth.getTransactionReceipt(txHash, function(err, receipt) {
-            if (err) {
-              return done(err);
-            }
-
-            assert.notEqual(receipt, null, "Transaction receipt shouldn't be null");
-            assert.equal(receipt.status, 0, "Reverted (failed) transactions should have a status of 0.");
-            return done();
-          });
-        });
-      }
-    });
+    assert.notStrictEqual(receipt, null, "Transaction receipt shouldn't be null");
+    assert.strictEqual(receipt.status, false, "Reverted (failed) transactions should have a status of FALSE.");
   });
 });
